% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{add_to_list_if_exists}
\alias{add_to_list_if_exists}
\title{Add Object to List if Exists}
\usage{
add_to_list_if_exists(list_obj, obj_name, var_name)
}
\arguments{
\item{list_obj}{A list to which the variable should be added if it exists.}

\item{obj_name}{A character string indicating the name to assign to the object within the list.}

\item{var_name}{A character string indicating the name of the variable to check for in the environment.}
}
\value{
The updated list with the variable added if it exists; otherwise, returns the list unchanged.
}
\description{
Checks if a variable exists in the environment and, if so, retrieves it,
adds it to the specified list with a given name, and logs a message about its creation.
}
\examples{
my_list <- list()
my_list <- add_to_list_if_exists(my_list, "example_obj", "example_var")
}
