% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{addAlpha}
\alias{addAlpha}
\title{Adds an alpha value to colors.}
\usage{
addAlpha(col, alpha = 0.1)
}
\arguments{
\item{A}{character vector of colors.}

\item{numeric}{alpha value between 0 and 1. Default is \code{.1}}
}
\value{
A vector of colors adjusted with the argument alpha value.
}
\description{
Adds an alpha value to colors.
}
