% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{addGeneIdsCols2HomologsKaKsTable}
\alias{addGeneIdsCols2HomologsKaKsTable}
\title{Extracts the gene pairs' identifiers from the 'Sequence' column of
'hom.kaks.tbl' and replaces this column with the parse result.}
\usage{
addGeneIdsCols2HomologsKaKsTable(hom.kaks.tbl)
}
\arguments{
\item{hom.kaks.tbl}{the result of calling 'readKaKsCalculatorOutput(...)' on
the result file of KaKs_Calculator used on homologs, that were NOT conserved
orthologs.}
}
\value{
A modified version of 'hom.kaks.tbl' in which the column 'Sequence'
is replaced with two columns 'gene.1' and 'gene.2'.
}
\description{
Extracts the gene pairs' identifiers from the 'Sequence' column of
'hom.kaks.tbl' and replaces this column with the parse result.
}
