% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{addGeneIdsCols2OrthologsKaKsTable}
\alias{addGeneIdsCols2OrthologsKaKsTable}
\title{Uses 'gene.1' as the first member of the orthologous gene pair and extracts
the gene identifier of the pairs second member from the 'Sequence' column of
'orth.kaks.tbl' and replaces this column with the result.}
\usage{
addGeneIdsCols2OrthologsKaKsTable(orth.kaks.tbl, gene.1)
}
\arguments{
\item{orth.kaks.tbl}{the result of calling 'readKaKsCalculatorOutput(...)'
on the result file of KaKs_Calculator used on conserved orthologs.}

\item{gene.1}{the identifier of the anchoring gene used within the currently
processed group of conserved orthologs.}
}
\value{
A modified version of 'orth.kaks.tbl' in which the column 'Sequence'
is replaced with two columns 'gene.1' and 'gene.2'.
}
\description{
Uses 'gene.1' as the first member of the orthologous gene pair and extracts
the gene identifier of the pairs second member from the 'Sequence' column of
'orth.kaks.tbl' and replaces this column with the result.
}
