% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{alignAAStringSet}
\alias{alignAAStringSet}
\title{Aligns the set of amino acid sequences based on the chemical properties of
their respective amino acid residues. Does so using MAFFT. NOTE: mafft is
expected to be installed.}
\usage{
alignAAStringSet(path.2.aa.seqs, path.2.aa.msa,
  mafft.call = getOption("mafft.call", paste("mafft --thread",
  getOption("mc.cores", 1), "--auto")))
}
\arguments{
\item{'path.2.aa.seqs'}{the valid file path to the amino acid sequences
FASTA}

\item{'path.2.aa.msa'}{the valid file path to the amino acid multiple
sequence alignment file that is to be generated}

\item{'mafft.call'}{the string used to build the system command which
invokes MAFFT.}
}
\value{
TRUE if and only of no error occurred.
}
\description{
Aligns the set of amino acid sequences based on the chemical properties of
their respective amino acid residues. Does so using MAFFT. NOTE: mafft is
expected to be installed.
}
