% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{alignCDSSetWithAlignedAAsAsGuide}
\alias{alignCDSSetWithAlignedAAsAsGuide}
\title{Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.}
\usage{
alignCDSSetWithAlignedAAsAsGuide(unaligned.cds.set, aligned.aa.set)
}
\arguments{
\item{'unaligned.cds.set'}{an instance of \code{base::list} as generated by
\code{seqinr::read.fasta} representing the unaligned codon sequences}

\item{'aligned.aa.set'}{an instance of \code{base::list} as generated by
\code{seqinr::read.fasta} representing the aligned amino acid sequences used
as guide}
}
\value{
An instance of \code{base::list} with names being the gene
accessions and values Strings representing the aligned coding-sequences.
}
\description{
Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.
}
