% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{alignCDSWithAlignedAASeq}
\alias{alignCDSWithAlignedAASeq}
\title{Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.}
\usage{
alignCDSWithAlignedAASeq(cds, aligned.aa.seq, aa.gap = "-",
  codon.gap = "---")
}
\arguments{
\item{'cds'}{a String representing the codon sequence to align.}

\item{'aligned.aa.seq'}{a String representing the aligned amino acid
sequence.}

\item{'aa.gap'}{the string used to represent a gap in the aligned AA sequence}

\item{'codon.gap'}{the string used to represent a gap in the aligned codon
sequence}
}
\value{
A String representing the aligned codon sequence
}
\description{
Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.
}
