% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{alignCodingSequencesPipeline}
\alias{alignCodingSequencesPipeline}
\title{Runs the pipeline to align a set of coding sequences: First translates them,
then validates them for premature stop codons, subsequently generates a
multiple sequence alignment (MSA) of amino acid (AA) sequences, then uses
this AA MSA as guide and aligns the coding sequences in the final step.}
\usage{
alignCodingSequencesPipeline(cds, work.dir, gene.group.name)
}
\arguments{
\item{cds}{an instance of \code{base::list} as generated by
\code{seqinr::read.fasta} representing the coding sequences that need to be
aligned}

\item{work.dir}{the working directory to use and in which to save the
relevant files}

\item{gene.group.name}{a string being used to name the output files written
into work.dir. Could be something like 'fam1234'.}
}
\value{
The ALIGNED and validated coding sequences as an instance of
\code{base::list} as generated by \code{seqinr::read.fasta}, or nothing if
validation discards the rest of 'cds'.
Sanitize the gene identifiers:
Convert to AA and align the AA-sequences:
Remove invalid AA-Sequences, i.e. AA-Seqs with premature stop-codons:
Warn about removed AA-Seqs:
If only a single sequence is left, we're done:
Write out the sanitized amino acid seqs:
Generate a multiple sequence alignment:
Use the aligned AA-Seqs as quide to align the CDS Sequences:
Return the CDS MSA using the ORIGINAL gene identifiers:
}
\description{
Runs the pipeline to align a set of coding sequences: First translates them,
then validates them for premature stop codons, subsequently generates a
multiple sequence alignment (MSA) of amino acid (AA) sequences, then uses
this AA MSA as guide and aligns the coding sequences in the final step.
}
