% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{alignedForUnalignedAAPos}
\alias{alignedForUnalignedAAPos}
\title{Computes the position of the unaligned character at argument position
'char.pos' in the aligned sequence of argument 'gene'.}
\usage{
alignedForUnalignedAAPos(gene, char.pos, msa.fasta,
  gap.char = getOption("GeneFamilies.gap.char", "-"))
}
\arguments{
\item{gene}{The gene accession / ID as used in 'msa.fasta'}

\item{char.pos}{The index of the homologous amino acid subject to selection
(integer coordinate)}

\item{msa.fasta}{The result of
\code{seqinr::read.fasta(path_2_AAs_MSA.fasta, seqtype='AA', as.string=TRUE,
strip.desc=TRUE)}.}

\item{gap.char}{The character used to identify non sequence characters in
the aligned sequences. Default is \code{getOption('GeneFamilies.gap.char',
'-')}.}
}
\value{
An integer; either NA if the position is a non sequence character,
or the corresponding aligned position.
}
\description{
Computes the position of the unaligned character at argument position
'char.pos' in the aligned sequence of argument 'gene'.
}
