% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{annotationBasedShannonEntropy}
\alias{annotationBasedShannonEntropy}
\title{Compute Shannon Entropy (\code{\link[entropy]{entropy}}) as a measure of
annotation (function) diversity within a given set of genes. If gene(s) have
multiple annotations these are sorted and concatonated in order to treat
them as single composite annotations. This means, that e.g if all genes have
the same three annotations the resulting entropy will be 0.}
\usage{
annotationBasedShannonEntropy(gene.accs, gene.annos = all.ipr, gene.col = 1,
  anno.col = 2,
  entropy.funk = getOption("MaizeGeneFamilies.entropy.function",
  entropy::entropy))
}
\arguments{
\item{gene.accs}{The identifiers or accessions of the genes to compute the
entropy for}

\item{gene.annos}{The data.frame holding the annotations for the genes in
'gene.accs'. Default is all available InterPro annotations expected to be
found in 'all.ipr'}

\item{gene.col}{the column of 'gene.annos' in which to lookup the gene
identifiers or gene accessions. Default is 1}

\item{anno.col}{the column of 'gene.annos' in which to lookup the function
annotation for the genes in 'gene.accs'. Default is 2}

\item{entropy.funk}{The function to be used to compute the Shannon Entropy.
Default is \code{getOption( 'MaizeGeneFamilies.entropy.function',
entropy::entropy )}. Set to \code{MaizeGeneFamilies::shannonEntropy} if you want
normalized Shannon Entropy. See
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4696313/}{Entropy
Explorer R package}.}
}
\value{
A list with two entries: 'entropy' is the computed shannon entropy,
and 'gene.annos.freqs' is an instance of table showing the annotation
architectures and their frequencies
}
\description{
Compute Shannon Entropy (\code{\link[entropy]{entropy}}) as a measure of
annotation (function) diversity within a given set of genes. If gene(s) have
multiple annotations these are sorted and concatonated in order to treat
them as single composite annotations. This means, that e.g if all genes have
the same three annotations the resulting entropy will be 0.
}

