% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{annotationEvolution}
\alias{annotationEvolution}
\title{Measures annotations gained, lost, and shared between base genes and evolved
genes.}
\usage{
annotationEvolution(genes, gene.sets, base.gene.set = "ortholog",
  annot.df = all.ipr, gene.col = 1, annot.col = 2)
}
\arguments{
\item{genes}{A group of genes in the form of a character vector of gene IDs}

\item{gene.sets}{A named list defining base gene and evolved gene
identifiers.}

\item{base.gene.set}{The name of list-entry in \code{gene.sets} indicating
which is to be interpreted as base genes. Default is \code{'ortholog'}.}

\item{annot.df}{A data.frame of gene annotations. Default is
\code{GeneFamilies::all.ipr}}

\item{gene.col}{The column name or index in \code{annot.df} in which to
lookup the gene IDs. Default is \code{1}}

\item{anno.col}{The column name or index in \code{annot.df} in which to
lookup the gene annotations. Default is \code{2}.}
}
\value{
A data.frame with the following columns: n.gained, n.lost, n.shared,
and total.
}
\description{
Measures annotations gained, lost, and shared between base genes and evolved
genes.
}

