% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{annotationMatrix}
\alias{annotationMatrix}
\title{Generate a 2x2 contingency table for statistical hypothesis testing of over-
or underrepresentation of 'anno' in the case 'anno.tbl' when compared with
the whole universe 'universe.tbl'. Please note, that 'anno.tbl' is expected
to be a true subset of 'universe.tbl.'}
\usage{
annotationMatrix(anno, anno.tbl, universe.tbl, anno.tbl.anno.col = 1,
  universe.tbl.anno.col = 1, dim.names = list(c("annotated",
  "not.annotated"), c("case", "universe")))
}
\arguments{
\item{anno}{string representing the annotation, e.g. InterPro entry}

\item{anno.tbl}{annotation table: column 1 holds gene identifiers, and
column 2 annotations, e.g. InterPro entries. 'anno.tbl' represents the
'case'}

\item{universe.tbl}{annotation table, true superset of 'anno.tbl'.
Represents the 'universe' to test alternative hypotheses against.}

\item{anno.tbl.anno.col}{The column of 'anno.tbl' in which to lookup the
case gene identifiers. Default is 1.}

\item{universe.tbl.anno.col}{The column of 'universe.tbl' in which to lookup
the universe gene identifiers. Default is 1.}

\item{dim.names}{used to set the row and column names of the generated
contingency table. See base::matrix(...,dimnames) for more details. Default
value is 'dim.names=list( c( 'annotated', 'not.annotated' ), c( 'case',
'universe' ) )'}
}
\value{
An instance of base::matrix representing the contingency table
}
\description{
Generate a 2x2 contingency table for statistical hypothesis testing of over-
or underrepresentation of 'anno' in the case 'anno.tbl' when compared with
the whole universe 'universe.tbl'. Please note, that 'anno.tbl' is expected
to be a true subset of 'universe.tbl.'
}
