% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFunks.R
\name{annotationProfile}
\alias{annotationProfile}
\title{Generate a profile of gene annotations. Count all annotations that are
groupable by a regular expression. This function has been designed for
MapMan-Bins.}
\usage{
annotationProfile(anno.df = maize.mapMan, ad.i = 1:nrow(anno.df),
  anno.regex = "^\\\\d+", ad.anno.col = "BINCODE", ad.anno.name = "NAME")
}
\arguments{
\item{anno.df}{An instance of \code{base::data.frame} with at least one
column, holding the annotation codes. Default is \code{maize.mapMan}.}

\item{ad.i}{logical or integer vector indicating which subset of argument
\code{anno.df} to consider. Default is to consider it completely,
\code{1:nrow(anno.df)}.}

\item{anno.regex}{A regular expression to identify groups of annotations.
Default is \code{'^\\d+\\.\\d+$'}.}

\item{ad.anno.col}{The column index or name of argument \code{anno.df} in
which to find the annotation codes. Default is \code{'BINCODE'}.}

\item{ad.anno.col}{The column index or name of argument \code{anno.df} in
which to find the annotation name. Default is \code{'NAME'}.}
}
\value{
An instance of \code{base::data.frame} with the following columns:
'Anno.Group' holds the group of annotation, 'Count' the number of matching
annotations, and 'Name' holds the annotations' names.
}
\description{
Generate a profile of gene annotations. Count all annotations that are
groupable by a regular expression. This function has been designed for
MapMan-Bins.
}

