% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{appendToTestedHypotheses}
\alias{appendToTestedHypotheses}
\title{Function appends a row to the argument data.frame 'df', which is supposed to
hold two columns 'tested.hypothesis' and 'p.val'.}
\usage{
appendToTestedHypotheses(df, hypo, p.val)
}
\arguments{
\item{df}{an instance of class base::data.frame with two columns}

\item{hypo}{string describing the hypothesis tested}

\item{p.val}{numerical result of testing 'hypo' as alternative hypothesis.}
}
\value{
The extended data.frame 'df' with a new row holding the arguments
'hypo' and 'p.val'.
}
\description{
Function appends a row to the argument data.frame 'df', which is supposed to
hold two columns 'tested.hypothesis' and 'p.val'.
}
