% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{areNeighbors}
\alias{areNeighbors}
\title{Informs whether two genes are located on the same chromosome and are within
a maximum distance. Distance is measured NOT in nucleotides, but in number
of separating genes.}
\usage{
areNeighbors(gene.a, gene.b, neighborhood.tbl,
  gene.col = getOption("GeneFamilies.neighborhood.gene.col", 2),
  chromosome.col = getOption("GeneFamilies.neighborhood.chromosome.col", 1),
  max.dist = getOption("GeneFamilies.neighborhood.max.dist", 9))
}
\arguments{
\item{gene.a}{The gene identifier for the first gene in question}

\item{gene.b}{The gene identifier for the second gene in question}

\item{neighborhood.tbl}{an instance of \code{base::data.frame} with at least
two columns. One holding the gene identifiers, and the other the chromosome
name. This table MUST be sorted by chromose and gene position, i.e. genes
separated by three rows and being located on the same chromosome are
expected to be separated by three genes on the respective chromosome.}

\item{gene.col}{Column name or number of \code{neighborhood.tbl} indicating
where to find the gene identifier. Default is
\code{getOption('GeneFamilies.neighborhood.gene.col', 2)}.}

\item{chromosome.col}{Column name or number of \code{neighborhood.tbl}
indicating where to find the chromosome names. Default is
\code{getOption('GeneFamilies.neighborhood.chromosome.col', 1)}.}

\item{max.dist}{An integer defining the maximum number of genes allowed to
separate two genes to be still considered 'neighbors'. Default is
\code{getOption('GeneFamilies.neighborhood.max.dist', 9)}}
}
\value{
A boolean or NA, if the two genes are not found in
\code{neighborhood.tbl}.
}
\description{
Informs whether two genes are located on the same chromosome and are within
a maximum distance. Distance is measured NOT in nucleotides, but in number
of separating genes.
}
