% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{axt}
\alias{axt}
\title{Extracts pairs of genes from a CDS multiple sequence alignment.}
\usage{
axt(cds.msa, anchor.gene)
}
\arguments{
\item{cds.msa}{an instance of Biostrings::DNAStringSet representing the
multiple alignment of coding sequences.}

\item{anchor.gene}{the gene identifier for which to build pairs with the
other genes present in 'cds.msa'.}
}
\value{
A named list. Names are the non anchor gene identifier that are
paired with the anchor gene. Values are the gene pairs themselfes.
}
\description{
Extracts pairs of genes from a CDS multiple sequence alignment.
}
