% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{buildPhylogeneticTree}
\alias{buildPhylogeneticTree}
\title{Uses FastTree[MP] to generate a maximum likelihood tree based on sequence
profiles.}
\usage{
buildPhylogeneticTree(path.2.cds.msa, path.2.ml.tree,
  fast.tree.call = getOption("fast.tree.call", paste("OMP_NUM_THREADS=",
  getOption("mc.cores", 1), " FastTreeMP -nt -gtr -gamma", sep = "")))
}
\arguments{
\item{'path.2.cds.msa'}{a valid file path to the codon multiple sequence
alignment to feed into FastTree[MP]}

\item{'path.2.ml.tree'}{a valid file path in which to store the resulting
tree in newick format}

\item{'fast.tree.call'}{the command whith which to invoke FastTree[MP]}
}
\value{
TRUE if and only if no error occurred.
}
\description{
Uses FastTree[MP] to generate a maximum likelihood tree based on sequence
profiles.
}
