% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{calculate_exp.prof.dists.statistics}
\alias{calculate_exp.prof.dists.statistics}
\title{Calculate Statistics for Expression Profile Distances}
\usage{
calculate_exp.prof.dists.statistics(data)
}
\arguments{
\item{data}{A list of matrices where each element represents a gene family's expression profile distance matrix.}
}
\value{
A tibble with columns: \code{Family}, \code{Mean}, and \code{Median}, containing the mean and median
values of the distances for each family.
}
\description{
Computes the mean and median of the expression profile distances for each gene family
within a given dataset.
}
\examples{
result <- calculate_exp.prof.dists.statistics(my_data)
}
