% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{calculate_exp.prof.dists.tissue.statistics}
\alias{calculate_exp.prof.dists.tissue.statistics}
\title{Calculate Statistics for Expression Profile Distances by Tissue}
\usage{
calculate_exp.prof.dists.tissue.statistics(data)
}
\arguments{
\item{data}{A nested list where each element represents a gene cluster, and each cluster
contains a list of tissues with associated expression profile distances.}
}
\value{
A tibble where each row represents a gene cluster, and columns include \code{Cluster} and
\code{Tissue-specific} mean and median values.
}
\description{
Computes the mean and median of the expression profile distances for each gene cluster
per tissue within a given dataset, transforming the output to have tissues in headers.
}
\examples{
result <- calculate_exp.prof.dists.tissue.statistics(my_data)
}
