% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{cbindAnnotationDescription}
\alias{cbindAnnotationDescription}
\title{Adds a column to 'enrichment.tbl' holding human readable descriptions for
the annotations found o be enriched. This function has been written
primarily for use with InterPro annotations.}
\usage{
cbindAnnotationDescription(enrichment.tbl, annotation.db = ipr.db,
  ano.desc.field = "NAME")
}
\arguments{
\item{enrichment.tbl}{a data.frame result of function 'enrichmentTests' or
'selectSignificantEnrichments'}

\item{annotation.db}{a list in which the names are function annotation
identifiers, e.g. IPR666666}

\item{ano.desc.field}{the name of the list entry to select in order to
obtain the description from members of 'annotation.db'}
}
\value{
A copy of 'enrichment.tbl' with an additional column holding the
decriptions of the respective annotations.
}
\description{
Adds a column to 'enrichment.tbl' holding human readable descriptions for
the annotations found o be enriched. This function has been written
primarily for use with InterPro annotations.
}
