% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{classSpecificExpressionProfileDists}
\alias{classSpecificExpressionProfileDists}
\title{Computes the within \code{stats} of expression profile distances for classes
of gene groups separating the argument genes into sub-groups.}
\usage{
classSpecificExpressionProfileDists(e.p.d, gene.classes, fam.name = NA,
  stats = c("max", "median", "mean", "maxMinusMin"))
}
\arguments{
\item{e.p.d}{An instance of \code{stats::dist} holding the pairwise
expression profile distances for the argument genes.}

\item{gene.classes}{An instance of \code{list} where the names are the
different gene classes and the values character vectors intersecting with
the argument genes and thus segregating them into the respective
gene-classes.}

\item{fam.name}{A string to be assigned as \code{Family} column in the
\code{base::data.frame} in slot \code{stats} of the returned object. Default
is \code{NA}.}

\item{stats}{A character vector holding the names of the functions to apply
on the the gene.class specific distances. Default is \code{stats = c('max',
'median', 'mean', 'maxMinusMin')}.}
}
\value{
An instance of \code{base::list} with one \code{base::dist} per
Gene.Class in \code{names(gene.classes)} and an additional
\code{base::data.frame} named \code{'stats'}. The latter holds the values
returned by the functions in \code{stats} applied to each separate class of
argument genes.
}
\description{
Computes the within \code{stats} of expression profile distances for classes
of gene groups separating the argument genes into sub-groups.
}
