% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{closestHomolog}
\alias{closestHomolog}
\title{Uses pairwise sequence similarity search results in order to identify which
gene is the closest homolog of the argument \code{gene.id}. Self-Matches are
excluded and as measure of sequence similarity the Bit-Score is used.}
\usage{
closestHomolog(gene.id, seq.sim.tbl = if (exists("all.vs.all.sim"))
  all.vs.all.sim, gene.col.a = "V1", gene.col.b = "V2",
  bit.score.col = "V12")
}
\arguments{
\item{gene.id}{The accession of the gene for which to find it's closest
homolog.}

\item{seq.sim.tbl}{An instance of \code{base::data.frame} holding the
tabular output of the pairwise sequence similarity searches. Could be
generated by using BLAT or Blast. Default is \code{if
(exists('all.vs.all.sim')) all.vs.all.sim}.}

\item{gene.col.a}{The column-name of \code{seq.sim.tbl} in which to find
argument \code{gene.id}. MUST be a string not a number. Default is
\code{'V1'}}

\item{gene.col.b}{The column-name of \code{seq.sim.tbl} in which to find
gene accessions ('hits') of significant similarity to the argument gene
\code{gene.id}. MUST be a string not a number. Default is \code{'V2'}}

\item{bit.score.col}{The column name of \code{seq.sim.tbl} in which to find
the numeric Bit-Scores of the pairwise local sequence alignments, generated
by Blast or BLAT e.g. MUST be a name not a number. Default is \code{'V12'}.}
}
\value{
The gene identifier (accession) of that gene that has the highest
Bit-Score in the pairwise comparisons with the argument gene \code{gene.id}.
If no such gene can be found NA is returned.
}
\description{
Uses pairwise sequence similarity search results in order to identify which
gene is the closest homolog of the argument \code{gene.id}. Self-Matches are
excluded and as measure of sequence similarity the Bit-Score is used.
}
