% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseKaKs.R
\docType{data}
\name{codeml.tmpl}
\alias{codeml.tmpl}
\title{Template for brew to generate the control file used as input for PAML's
codeml.}
\format{An object of class \code{character} of length 1.}
\usage{
codeml.tmpl
}
\description{
Template for brew to generate the control file used as input for PAML's
codeml.
}
\keyword{datasets}
