% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{compositeAnnotations}
\alias{compositeAnnotations}
\title{Generates a list of composite Annotations found for the argument genes based
on the argument gene (function) annotation data.frame. Composite annotations
are sorted alphabetically and then concatonated with ',' as separator.}
\usage{
compositeAnnotations(gene.accs, gene.annos = all.ipr, gene.col = 1,
  anno.col = 2)
}
\arguments{
\item{gene.accs}{The identifiers or accessions of the genes to compute the
entropy for}

\item{gene.annos}{The data.frame holding the annotations for the genes in
'gene.accs'. Default is all available InterPro annotations expected to be
found in 'all.ipr'}

\item{gene.col}{the column of 'gene.annos' in which to lookup the gene
identifiers or gene accessions. Default is 1}

\item{anno.col}{the column of 'gene.annos' in which to lookup the function
annotation for the genes in 'gene.accs'. Default is 2}
}
\value{
An character vector with names being the gene identifiers
'gene.accs' and values the composite (function) annotations.
}
\description{
Generates a list of composite Annotations found for the argument genes based
on the argument gene (function) annotation data.frame. Composite annotations
are sorted alphabetically and then concatonated with ',' as separator.
}
