% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{computeOrthologSpecificAnnotationDiversityPerGroup}
\alias{computeOrthologSpecificAnnotationDiversityPerGroup}
\title{Wrapper for function
\code{computeSpeciesSpecificAnnotationDiversityPerGroup}. Instead of species
affiliation the distinction here is wether genes are Orthologs or
Non-Orthologs.}
\usage{
computeOrthologSpecificAnnotationDiversityPerGroup(gene.groups.df,
  orthologs = orthologs.genes, group.col = "Family",
  gene.col = "Gene")
}
\arguments{
\item{gene.groups.df}{An instance of base::data.frame with at least two
columns. One must hold the gene group identifiers and the other the gene
identifiers (accessions) belonging to the respective group.}

\item{group.col}{The column name of 'gene.groups.df' in which to find the
gene group identifiers. Default is 'Family'.}

\item{gene.col}{The column name of 'gene.groups.df' in which to find the
gene identifiers (accessions). Default is 'Gene'.}
}
\value{
An instance of base::data.frame with several columns. The first
holds the gene group identifiers obtained from 'gene.groups.df' and one
numeric column for each \code{Orthologs} and \code{Non-Orthologs}. The later
columns hold the gene-class specific annotation entropies for the respective
gene groups; NA values where no genes of that species were found to be
members of the respective gene group.
}
\description{
Wrapper for function
\code{computeSpeciesSpecificAnnotationDiversityPerGroup}. Instead of species
affiliation the distinction here is wether genes are Orthologs or
Non-Orthologs.
}
