% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{conservedOrthologs}
\alias{conservedOrthologs}
\title{Identifies those pairs of orthologous genes where one member appears in all
pairwise species comparisons.}
\usage{
conservedOrthologs(pairwise.orths, intersect.col = "V1",
  ortholog.col = "V2", anchor.species = "chi")
}
\arguments{
\item{pairwise.orths}{a list of two column data frames holding orthologous
gene pairs identified in pairwise species comparisons. The names of this
list reflect the species the anchoring species was compared to iteratively}

\item{intersect.col}{the name or index of the column of each of the data
frames in 'pairwise.orths' to use as a source of gene identifiers}

\item{ortholog.col}{the name or index of the column to extract ortholog
genes from when they match anchor species orthologs conserved over all
species comparisons}

\item{anchor.species}{the name of the species which was used in all pairwise
comparisons.}
}
\value{
a two column data frame of orthologous gene groups where an ortholog
was found in each species.
}
\description{
Identifies those pairs of orthologous genes where one member appears in all
pairwise species comparisons.
}
