% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{constructAnnotationVectorSpace}
\alias{constructAnnotationVectorSpace}
\title{For the argument \code{genes} construct a vector space based on the distinct
annotations found for the respective genes in arument \code{annot.df}. Each
distinct annotation will be an axis, each gene will be assigned a vector,
with coordinates the discrete number of times an annotation has been
assigned to the respective gene.}
\usage{
constructAnnotationVectorSpace(genes, annot.df = all.ipr, gene.col = 1,
  annot.col = 2)
}
\arguments{
\item{genes}{A character vector of gene IDs}

\item{annot.df}{A data.frame or matrix holding the gene annotations. Default
is \code{all.ipr}.}

\item{gene.col}{The column name or number of \code{annot.df} in which to
look up the gene IDs found in \code{genes}. Default is \code{1}.}

\item{annot.col}{The column name or number of \code{annot.df} in which to
look up the genes annotations. Default is \code{2}.}
}
\value{
\code{NA} if no annotations are found for any gene in \code{genes},
otherwise a data.frame with a column \code{gene} in which the gene IDs are
held and one column for each distinct annotation or the respective vector
space axes.
}
\description{
For the argument \code{genes} construct a vector space based on the distinct
annotations found for the respective genes in arument \code{annot.df}. Each
distinct annotation will be an axis, each gene will be assigned a vector,
with coordinates the discrete number of times an annotation has been
assigned to the respective gene.
}
