% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{cosAngleVec}
\alias{cosAngleVec}
\title{Computes the cosine of the angle between the two n-dimensional argument
vectors \code{a} and \code{b}.}
\usage{
cosAngleVec(a, b)
}
\arguments{
\item{a}{numerical representing first vector}

\item{b}{numerical representing second vector}
}
\value{
cosine of angle between the two argument vectors or NaN if any entry
in the vectors is not a number.
}
\description{
Computes the cosine of the angle between the two n-dimensional argument
vectors \code{a} and \code{b}.
}
