% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{cosDiag}
\alias{cosDiag}
\title{Compute cosinus of the angle between argument vector \code{x} and the vector
space diagonal.}
\usage{
cosDiag(x, d.v = rep(1, length(x)))
}
\arguments{
\item{x}{numeric representing a vector in n-dimensional space}

\item{d.v}{numeric the diagonal vector in n-dimensional space. Default is
\code{rep(1, length(x))}.}
}
\value{
Returns the cosinus of the angle between argument vector \code{x}
and the vector space diagonal. If any value in argument \code{x} is infinite
or not a number or if \code{x} is \code{NULL} \code{NaN} is returned.
}
\description{
Compute cosinus of the angle between argument vector \code{x} and the vector
space diagonal.
}
