% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{countNonAminoAcidCharsFromStart}
\alias{countNonAminoAcidCharsFromStart}
\title{Counts the number of non amino acid characters at the start of the
argument 'aa.seq'.}
\usage{
countNonAminoAcidCharsFromStart(aa.seq, non.aa.regex = "^[-]+")
}
\arguments{
\item{aa.seq}{The aligned amino acid sequence possibly including gap
characters}

\item{non.aa.regex}{The PERL regular expression to identify non amino acid
characters at the start of the argument 'aa.seq'}
}
\value{
An integer - the count of non amino acid chars at the beginning
of the argument 'aa.seq'.
}
\description{
Args:
}
