% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{distNormExprVectrs}
\alias{distNormExprVectrs}
\title{Returns the vector resulting from subtracting \code{e.v.b} from
\code{e.v.a}. Both argument vectors are normalized before subtraction.}
\usage{
distNormExprVectrs(e.v.a, e.v.b)
}
\arguments{
\item{e.v.a}{numeric}

\item{e.v.b}{numeric}
}
\value{
\code{NA} if any of the arguments contains components that are non
legal coordinates (see \code{hasInvalidVecSpaceComponents}). Otherwise
\code{normalizeVector(e.v.a) - normalizeVector(e.v.b)} is returned.
}
\description{
Returns the vector resulting from subtracting \code{e.v.b} from
\code{e.v.a}. Both argument vectors are normalized before subtraction.
}
