% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{distVectorClouds}
\alias{distVectorClouds}
\title{Measures the distance between two vector clouds (see
\code{GeneFamilies::statVectorCloud}) within the same vector space. The
vector between the two \code{stat.vec}s representing \code{cl.a} and
\code{cl.b} is computed. The deviance spaces of the two respective clouds
are projected onto this distance vector, and the remaining 'un-shadowed'
length returned. If this length is positive the two clouds do not overlap
considering their representatives and the irespective deviance spreads
around them.}
\usage{
distVectorClouds(cl.a, cl.b)
}
\arguments{
\item{cl.a}{Result of invoking \code{GeneFamilies::statVectorCloud} on a set
of vectors.}

\item{cl.b}{Result of invoking \code{GeneFamilies::statVectorCloud} on a set
of vectors.}
}
\value{
Numeric - The length of the difference vector being outside the
respective clouds' deviances.
}
\description{
Measures the distance between two vector clouds (see
\code{GeneFamilies::statVectorCloud}) within the same vector space. The
vector between the two \code{stat.vec}s representing \code{cl.a} and
\code{cl.b} is computed. The deviance spaces of the two respective clouds
are projected onto this distance vector, and the remaining 'un-shadowed'
length returned. If this length is positive the two clouds do not overlap
considering their representatives and the irespective deviance spreads
around them.
}
