% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{domainsForPos}
\alias{domainsForPos}
\title{Looks up the conserved protein domains (InterPro) that have been annotated
to gene and overlap with amino acid position 'aa.pos'.}
\usage{
domainsForPos(gene, aa.pos, iprscan.tbl, gene.col = "V1",
  start.col = "V7", end.col = "V8", ipr.col = "V12")
}
\arguments{
\item{gene}{The gene accession or ID as used in iprscan.tbl}

\item{aa.pos}{The amino acid position (non aligned) to be overlapped by any
InterPros}

\item{gene.col}{The column index or name of iprscan.tbl in which to look up
the gene accessions}

\item{start.col}{The column index or name of iprscan.tbl in which to look up
the domain start positions}

\item{end.col}{The column index or name of iprscan.tbl in which to look up
the domain end positions}

\item{ipr.col}{The column index or name of iprscan.tbl in which to look up
the InterPro identifier}
}
\value{
A character vector of matching InterPro domains.
}
\description{
Looks up the conserved protein domains (InterPro) that have been annotated
to gene and overlap with amino acid position 'aa.pos'.
}
