% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{enrichedAnnotations}
\alias{enrichedAnnotations}
\title{Tests the NULL Hypothesis for each gene (function) annotation in
\code{annos.2.test} that its annotation-frequency in the \code{case.genes}
is explicable with the background distribution observed in
\code{universe.annos}. The default alternative hypothesis is that the actual
case frequency is 'greater'.}
\usage{
enrichedAnnotations(case.genes, universe.annos = all.ipr,
  univ.gene.col = 1, univ.anno.col = 2,
  annos.2.test = unique(universe.annos[which(universe.annos[,
  univ.gene.col] \%in\% case.genes), univ.anno.col]),
  alt.hypothesis = "greater", p.adjust.method = "BY")
}
\arguments{
\item{case.genes}{A character vector of gene identifiers defining the gene
group of interest for which to infer enriched annotations.}

\item{universe.annos}{A data.frame with at least two columns: 1. of gene
identifiers and another of (function) annotations. Default is this package's
data \code{all.ipr}.}

\item{univ.gene.col}{The column of \code{universe.annos} in which to lookup
gene identifier. Default is 1.}

\item{univ.anno.col}{The column of \code{universe.annos} in which to lookup
annotations. Default is 2.}

\item{annos.2.test}{The annotations for which to test enrichment. Default is
all annotations found for the \code{case.genes}.}

\item{alt.hypothesis}{The alternative hypotheses to test. Default is that
the case frequencies are significantly greater than the one observed in the
'universe'.}

\item{p.adjust.method}{The method used to adjust P-Values for multiple
hypothesis testing. Default is 'BY'. See \code{?p.adjust} for more details.}
}
\value{
A numeric vector of P-Values adjusted to multiple hypothesis
testing. Names are the \code{annos.2.test}.
}
\description{
Tests the NULL Hypothesis for each gene (function) annotation in
\code{annos.2.test} that its annotation-frequency in the \code{case.genes}
is explicable with the background distribution observed in
\code{universe.annos}. The default alternative hypothesis is that the actual
case frequency is 'greater'.
}
