% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{enrichmentTests}
\alias{enrichmentTests}
\title{Tests for every function annotation in 'case.anno.tbl' the alternative
hypothesis provided with 'test.dir'. Resulting p.values are adjusted for
multiple hypothesis testing, before being returned. Note, if applied on
large data sets, run in parallel on multiple cores and set
options('mc.cores'=...) accordingly.}
\usage{
enrichmentTests(case.anno.tbl, universe.anno.tbl, test.dir = "greater",
  p.adjust.method = "BY")
}
\arguments{
\item{case.anno.tbl}{the annotations for which the alternative hypothesis is
to be tested}

\item{universe.anno.tbl}{the 'universe' annotations to compare the case with}

\item{test.dir}{the direction of the alternative hypothesis. Default defines
it as 'greater'}

\item{p.adjust.method}{the method used to adjust resulting p.values for
multiple hypothesis testing. Default is 'BY', see base::p.adjust for more
details.}
}
\value{
A data.frame holding the annotation in column 'PFam', the adjusted
and un-adjusted P.Value in two additional columns.
}
\description{
Tests for every function annotation in 'case.anno.tbl' the alternative
hypothesis provided with 'test.dir'. Resulting p.values are adjusted for
multiple hypothesis testing, before being returned. Note, if applied on
large data sets, run in parallel on multiple cores and set
options('mc.cores'=...) accordingly.
}
