% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{euclNorm}
\alias{euclNorm}
\title{Computes the Euclidean norm (length) of vector \code{x}.}
\usage{
euclNorm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A scalar numeric value, the length of the vector \code{x}.
}
\description{
Computes the Euclidean norm (length) of vector \code{x}.
}
