% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{exp.prof.dists}
\alias{exp.prof.dists}
\title{Compute Euclidean Distances Between Gene Expression Profiles}
\usage{
\method{exp}{prof.dists}(
  gene.accessions,
  expression.profiles = rna.seq.exp.profils,
  expr.prof.gene.col = "FBpp_ID",
  tissues = setdiff(colnames(expression.profiles), c(expr.prof.gene.col, "Parent_FBgn",
    "Species")),
  dist.method = "euclidean"
)
}
\arguments{
\item{gene.accessions}{A list of gene accession identifiers for which distances should be computed.}

\item{expression.profiles}{A data frame or tibble containing gene expression data. This should include a column for gene IDs and additional columns for expression values. Default is \code{rna.seq.exp.profils}.}

\item{expr.prof.gene.col}{The column name in \code{expression.profiles} containing gene IDs. Default is \code{"FBpp_ID"}.}

\item{tissues}{A character vector of column names in \code{expression.profiles} that represent tissue-specific expression values. By default, includes all columns except \code{expr.prof.gene.col}, \code{"Parent_FBgn"}, and \code{"Species"}.}

\item{dist.method}{Distance calculation method. Default is \code{"euclidean"}. Other methods (e.g., \code{"manhattan"}, \code{"maximum"}, etc.) can be specified.}
}
\value{
A list of vectors, each representing the pairwise Euclidean distances between gene expression profiles for a species. Each vector contains the distances between the specified genes' expression profiles across the selected tissues. If there are fewer than two expression profiles for a species, \code{NA} is returned for that species.
}
\description{
Calculates pairwise Euclidean distances across gene expression profiles for a given set of genes, aggregated across all specified tissues. The distances are computed within each species, providing a measure of similarity in gene expression profiles.
}
\examples{
# Compute overall distances for specified genes across all tissues
distances <- exp.prof.dists(gene.accessions = list("gene1", "gene2", "gene3"))

}
