% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{expressionProfilesDists}
\alias{expressionProfilesDists}
\title{Computes the distance (\code{base::dist}) between the expression profiles of
the genes in argument \code{gene.accessions}.}
\usage{
expressionProfilesDists(gene.accessions,
  expression.profiles = rna.seq.exp.profils,
  expr.prof.gene.col = "gene",
  expr.prof.gene.variant.col = "gene.exp.var",
  tissues = setdiff(colnames(expression.profiles), c(expr.prof.gene.col,
  expr.prof.gene.variant.col)), dist.method = "euclidean",
  per.tissue = FALSE)
}
\arguments{
\item{gene.accessions}{A character vector of gene identifiers some of which
are expected to be present in \code{expression.profiles}}

\item{expression.profiles}{A data.frame with gene expression profiles.
Default is this package's \code{rna.seq.exp.profils} data.}

\item{expr.prof.gene.col}{A number or string identifying the column of
\code{expression.profiles} in which to lookup the gene accessions. Default
is \code{'gene'}.}

\item{expr.prof.gene.variant.col}{A number or string identifying the column
of \code{expression.profiles} in which to lookup the gene accessions
including their expression variant identifier (e.g. 'AT1G12345.3'). Default
is \code{gene.exp.var}.}

\item{tissues}{A character or integer indexing those columns of
\code{expression.profiles} in which the actual relative expression per
tissue is stored. In short these columns are the expression profiles.
Default is \code{setdiff(colnames(expression.profiles),
c(expr.prof.gene.col, expr.prof.gene.variant.col))}.}

\item{dist.method}{A string to pass as argument \code{method} to
\code{base::dist}. Default is \code{'euclidean'}.}

\item{per.tissue}{Logical indicating wether to compute distances in the
euclidean space created by the dimensions of all tissues, or if set to
\code{TRUE} computes the one-dimensional distances per tissue. Default is
\code{FALSE}.}
}
\value{
An instance of base::dist holding pairwise distances between the
argument genes expression profiles. If not at least two expression profiles
are found for the argument \code{gene.accessions} \code{NA} is returned. In
case the argument \code{per.tissue} is set to \code{TRUE} a list of
one-dimensional distances is returned; names of the list are the respective
tissues.
}
\description{
Computes the distance (\code{base::dist}) between the expression profiles of
the genes in argument \code{gene.accessions}.
}
