% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{exprVecSpaceEvolutionAfterDupl}
\alias{exprVecSpaceEvolutionAfterDupl}
\title{Based on an expression vector space infers the evolution of gene function
after duplication by comparing base genes and groups of evolved genes.}
\usage{
exprVecSpaceEvolutionAfterDupl(genes, family.name, classifier.lst,
  base.class = "ortholog", expr.vecs = rpkm.expr.profiles.df,
  gene.col = "gene", vec.space.axes = c("cotyledon", "seedling",
  "developing leaf", "flower stage 9", "flower stage 16"))
}
\arguments{
\item{genes}{character vector of gene identifiers}

\item{family.name}{The name of the gene group comprising argument \code{genes}}

\item{classifier.lst}{A named list of at least two entries. Entries should
be lists of gene IDs belonging to their respective (named) classes.}

\item{base.class}{A name of argument \code{classifier.lst} to be used as
base gene class. Default is \code{'ortholog'}.}

\item{expr.vecs}{A set of vectors in the respective expression vector space.
Each row holds the expression vector for its respective gene. Default is
\code{rpkm.expr.profiles.df}.}

\item{gene.col}{The name or index of a column of argument \code{expr.vecs}
indicating in which column to lookup gene IDs.}

\item{vec.space.axes}{A character or integer vector indicating the columns
of argument \code{expr.vecs} which correspond to the respective vector space
axes. Default is \code{c('cotyledon', 'seedling', 'developing leaf', 'flower
stage 9', 'flower stage 16')}.}
}
\value{
A data.frame with the following columns: Family (set to argument
\code{family.name}, mean.base.tiss.vers, and for each class 'EvCl' of
evolved genes EvCl.tiss.vers, EvCl.tiss.change, and EvCl.dist.vec.clouds
}
\description{
Based on an expression vector space infers the evolution of gene function
after duplication by comparing base genes and groups of evolved genes.
}
