% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{f}
\alias{f}
\title{ff ff f}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
f
}
\arguments{
\item{a}{A character string specifying the file path where the RData file will be saved.}
}
\value{
None. This function is used for its side effect of saving data to a file.
}
\description{
Saves the coding sequences of various Drosophila species as an RData file.
}
\examples{
f(a 
     file = file.path("data-drosophila/Genome-Fasta", "codingSequences-drosophila.RData"))

}
\keyword{datasets}
