% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{familyExpanded}
\alias{familyExpanded}
\title{Alternative method to detect expanded or compressed families comparing one
species with the background of the others. P-Values are obtained from an
outlier chisquare test; see package outliers::chisq.out.test(...) for
details.}
\usage{
familyExpanded(x, test.species = "chi",
  background.species = setdiff(names(x), test.species),
  expansion.test = max)
}
\arguments{
\item{x}{named integer vector representing the gene family to test. x holds
the number of gene members per species.}

\item{test.species}{a string being used ase a name in argument 'x'
indicating which species is the alternative hypothesis of expansion (or
contraction) based upon}

\item{background.species}{a character vector holding the names of the
species the background distribution of numbers of the species specific
family members is based upon.}

\item{expansion.test}{boolean indicating whether to test for expansion
(default TRUE) or compression (set to FALSE).}
}
\value{
numeric p-value obtained from outliers::chisq.out.test(x) or NA, if
expansion-test does not indicate a candidate outlier.
}
\description{
Alternative method to detect expanded or compressed families comparing one
species with the background of the others. P-Values are obtained from an
outlier chisquare test; see package outliers::chisq.out.test(...) for
details.
}
