% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_funks.R
\name{filter_gene_pairs}
\alias{filter_gene_pairs}
\title{Filter Gene Pairs Based on Expression Data}
\usage{
filter_gene_pairs(df, type = c("Ortholog", "Paralog"), expression_data)
}
\arguments{
\item{df}{A data frame containing gene pairs, with one column labeled \code{"Gene"} and another column containing either \code{"Ortholog"} or \code{"Paralog"} gene identifiers.}

\item{type}{A character string specifying the type of gene pair to filter: \code{"Ortholog"} or \code{"Paralog"}. Default is \code{"Ortholog"}.}

\item{expression_data}{A data frame or tibble containing expression data, with a column \code{"FBpp_ID"} listing gene IDs for which expression data is available.}
}
\value{
A filtered data frame containing only the gene pairs where both genes have corresponding entries in \code{expression_data}.
}
\description{
Filters gene pairs by retaining only those entries where both genes in each pair are present in the specified expression data. This reduces computation time by excluding pairs with no expression data.
}
\examples{
# Filter ortholog pairs based on available expression data
filtered_df <- filter_gene_pairs(df = gene_pairs, type = "Ortholog", expression_data = rna_seq_data)

# Filter paralog pairs based on available expression data
filtered_df <- filter_gene_pairs(df = gene_pairs, type = "Paralog", expression_data = rna_seq_data)

}
