% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_funks.R
\name{filter_v1}
\alias{filter_v1}
\title{Filter Genes Based on Expression Data}
\usage{
filter_v1(df, expression_data)
}
\arguments{
\item{df}{A data frame containing gene data, with a column labeled \code{"Gene"}.}

\item{expression_data}{A data frame or tibble containing expression data, with a column \code{"FBpp_ID"} listing gene IDs with available expression data.}
}
\value{
A filtered data frame containing only rows where the \code{"Gene"} column has matching entries in the \code{expression_data}.
}
\description{
Filters the input data frame by retaining only those rows where the gene in the \code{"Gene"} column has corresponding entries in the expression data. This function is intended to simplify the dataset by removing entries for genes with no available expression data.
}
\examples{
# Filter genes based on available expression data
filtered_df <- filter_v1(df = gene_data, expression_data = rna_seq_data)

}
