% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_funks.R
\name{filter_v2}
\alias{filter_v2}
\title{Filter Gene Pairs Based on Expression Data and Pair Type}
\usage{
filter_v2(df, type = c("Ortholog", "Paralog"), expression_data)
}
\arguments{
\item{df}{A data frame containing gene pairs, with one column labeled \code{"Gene"} and another column containing either \code{"Ortholog"} or \code{"Paralog"} gene identifiers, as specified by the \code{type} argument.}

\item{type}{A character string specifying the type of gene pair to filter by: \code{"Ortholog"} or \code{"Paralog"}. Default is \code{"Ortholog"}.}

\item{expression_data}{A data frame or tibble containing expression data, with a column \code{"FBpp_ID"} listing gene IDs for which expression data is available.}
}
\value{
A filtered data frame containing only the gene pairs where both the \code{"Gene"} and the specified \code{type} column have corresponding entries in \code{expression_data}.
}
\description{
Filters the input data frame by retaining only those rows where both the gene in the \code{"Gene"} column and its corresponding ortholog or paralog (as specified in the \code{type} parameter) have entries in the expression data. This function is useful for reducing datasets to only include pairs where both genes have expression data available.
}
\examples{
# Filter ortholog pairs based on available expression data
filtered_df <- filter_v2(df = gene_pairs, type = "Ortholog", expression_data = rna_seq_data)

# Filter paralog pairs based on available expression data
filtered_df <- filter_v2(df = gene_pairs, type = "Paralog", expression_data = rna_seq_data)

}
