% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{foregroundNodes}
\alias{foregroundNodes}
\title{Identifies those tip and inner nodes of the phylogenetic tree that span
subtrees whose tips are true subsets of the argument 'foreground.nodes'.}
\usage{
foregroundNodes(phylo.tr, foreground.nodes)
}
\arguments{
\item{phylo.tr}{instance of ape::phylo representing a phylogenetic tree}

\item{foreground.nodes}{a vector of tip indices or tip labels that represent
candidate foreground tips.}
}
\value{
integer vector of inner node indices that span subtrees containing
only foreground tips merged with the argument 'foreground.nodes'
If tip labels are passed as argument infer their index in the tree's nodes
slot:
Find those inner nodes that spawn subtrees whose tips are true subsets of
'foreground.nodes'
}
\description{
Identifies those tip and inner nodes of the phylogenetic tree that span
subtrees whose tips are true subsets of the argument 'foreground.nodes'.
}
