% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{formatNodeLabels}
\alias{formatNodeLabels}
\title{For a default phylogentic tree this formats the local support values for
printing.}
\usage{
formatNodeLabels(nd.lb)
}
\arguments{
\item{nd.lb}{the character vector of node labels of an instance of class
ape::phylo}
}
\value{
a formatted character vector to be used as a replacement for the
original node labels
}
\description{
For a default phylogentic tree this formats the local support values for
printing.
}
