% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{geneClassesVectorClouds}
\alias{geneClassesVectorClouds}
\title{Invokes \code{MaizeGeneFamilies::statVectorCloud} on each subset of genes in
\code{annot.vec.space} classified by the gene-sets in \code{gene.sets}.}
\usage{
geneClassesVectorClouds(annot.vec.space, gene.sets)
}
\arguments{
\item{annot.vec.space}{A data.frame result of calling function
\code{constructAnnotationVectorSpace}}

\item{gene.sets}{A named list, where names indicate gene-sets (or classes),
and values are gene IDs including the ones references in
\code{annot.vec.space}}
}
\value{
A named list, with one entry per gene-class in \code{gene.sets}.
Values are the results of invoking \code{MaizeGeneFamilies::statVectorCloud} on
each respective subset.
}
\description{
Invokes \code{MaizeGeneFamilies::statVectorCloud} on each subset of genes in
\code{annot.vec.space} classified by the gene-sets in \code{gene.sets}.
}

