% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{geneGroupsAnnotationBasedShannonEntropy}
\alias{geneGroupsAnnotationBasedShannonEntropy}
\title{Computes for each gene group in \code{groups.ids} the annotation based
Shannon Entropy in parallel. Set \code{options(mc.cores=detectCores())} for
maximum speed.}
\usage{
geneGroupsAnnotationBasedShannonEntropy(groups.ids, groups.2.genes.lst,
  gene.annos = all.ipr, gene.col = 1, anno.col = 2)
}
\arguments{
\item{groups.ids}{The IDs of the gene groups to compute the annotation based
Shannon Entropy for.}

\item{groups.2.genes.lst}{A list with names intersecting with argument
\code{groups.ids} and values character vectors holding gene-IDs found in
\code{gene.annos}.}

\item{gene.col}{the column of 'gene.annos' in which to lookup the gene
identifiers or gene accessions. Default is 1}

\item{anno.col}{the column of 'gene.annos' in which to lookup the function
annotation for the genes in 'gene.accs'. Default is 2}
}
\value{
A list with names \code{groups.ids} and values the computed
annotation based Shannon Entropies.
}
\description{
Computes for each gene group in \code{groups.ids} the annotation based
Shannon Entropy in parallel. Set \code{options(mc.cores=detectCores())} for
maximum speed.
}
