% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{geneGroupSubfunctionalization}
\alias{geneGroupSubfunctionalization}
\title{Invokes \code{subfunctionalizationAnnotationBased} for all gene.groups
(families) contained in argument \code{gene.groups.lst}, using argument
\code{gene.classes} to separate 'base genes' from 'evolved genes'.}
\usage{
geneGroupSubfunctionalization(gene.groups.lst, gene.classes)
}
\arguments{
\item{gene.groups.lst}{A named list of character vectors holding gene IDs.}

\item{gene.classes}{A named list of gene identifiers separated into a base
class and a class of evolved genes.}
}
\value{
NULL, if no subfunctionalization could be identified within any of
the gene families, or a data frame with the identified subfunctionalizations
(see \code{subfunctionalizationAnnotationBased} for details). An additional
column holds the name of the gene family.
}
\description{
Invokes \code{subfunctionalizationAnnotationBased} for all gene.groups
(families) contained in argument \code{gene.groups.lst}, using argument
\code{gene.classes} to separate 'base genes' from 'evolved genes'.
}
