% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{getAASubstring}
\alias{getAASubstring}
\title{Selects the aligned amino acid subsequence around the given coordinate
including radius up- and downstream AAs.}
\usage{
getAASubstring(msa.fasta, gene, sel.aa, radius = 3)
}
\arguments{
\item{msa.fasta}{The amino acid multiple sequence alignment (MSA) as
returned by readAAMultipleAlignment(…) [package Biostrings]}

\item{gene}{The gene identifier / accession as found in the MSA}

\item{sel.aa}{The index of the amino acid coordinate, starting at 1}

\item{radius}{The number of up- and downstream alignment characters to
return}
}
\value{
A list with two entries. Entry 'aa.subseq' is an instance of
AAString of max length 1+2*radius. If radius exceeds start or stop of the AA
sequence the result is shorter accordingly. Entry 'up.strm.offset' informs
about how many characters prepend the original selected amino acid according
to 'sel.aa'.
}
\description{
Selects the aligned amino acid subsequence around the given coordinate
including radius up- and downstream AAs.
}
