% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{getFamilyDNAStringSet}
\alias{getFamilyDNAStringSet}
\title{Extracts the coding sequences of the gene members of the argument family
from this package's database 'all.cds'. Warns, if not all members are found
in the database.}
\usage{
getFamilyDNAStringSet(fam, cds = all.cds)
}
\arguments{
\item{fam}{- The character vector of gene names member of the family.}

\item{cds}{- The database of coding sequences as an instance of \code{base::list}
[seqinr]. Default is this packages database 'all.cds'.}
}
\value{
An instance of \code{base::list} as generated by
\code{seqinr::read.fasta} holding the coding sequences of the argument
family.
}
\description{
Extracts the coding sequences of the gene members of the argument family
from this package's database 'all.cds'. Warns, if not all members are found
in the database.
}
