% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMcCores}
\alias{getMcCores}
\title{Reads out SHELL environment variable \code{MCCORES} and returns it as
integer. If the environment variable is not set, the value returned by
\code{detectCores()} is used as default.}
\usage{
getMcCores(env.var = "MCCORES")
}
\arguments{
\item{env.var}{The name of the SHELL environment variable to read out.
Default is \code{MCCORES}.}
}
\value{
An integer, either the value of the environment variable
\code{MCCORES} or the default value returned by
\code{parallel::detectCores()}.
}
\description{
Reads out SHELL environment variable \code{MCCORES} and returns it as
integer. If the environment variable is not set, the value returned by
\code{detectCores()} is used as default.
}
