% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{getPairsFromDistForIndices}
\alias{getPairsFromDistForIndices}
\title{Identifies the pairs that match the argument indices \code{k} in the
argument \code{dist.obj}. E.g. you might want to know which pairs have
maximum distance: \code{getPairsFromDistForIndices(which(x.dist ==
max(x.dist)), x.dist)} will help you out. Code taken from this post:
https://stackoverflow.com/questions/39005958/r-how-to-get-row-column-subscripts-of-matched-elements-from-a-distance-matri}
\usage{
getPairsFromDistForIndices(k, dist.obj)
}
\arguments{
\item{k}{an integer vector holding the indices for which to identify the
pairs.}

\item{dist.obj}{An instance of \code{base::dist} that holds the distances
and labels from which to extract the pairs.}
}
\value{
An instance of \code{base::matrix} with two columns in which each
row holds a pair. Cells can be \code{NA} if the argument indices were not
valid.
}
\description{
Identifies the pairs that match the argument indices \code{k} in the
argument \code{dist.obj}. E.g. you might want to know which pairs have
maximum distance: \code{getPairsFromDistForIndices(which(x.dist ==
max(x.dist)), x.dist)} will help you out. Code taken from this post:
https://stackoverflow.com/questions/39005958/r-how-to-get-row-column-subscripts-of-matched-elements-from-a-distance-matri
}
