% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{getSpeciesNodeNumberFromCafe}
\alias{getSpeciesNodeNumberFromCafe}
\title{Extracts the number CAFE assigned the species during its analysis.}
\usage{
getSpeciesNodeNumberFromCafe(cafe.id.line, species.name = "chi")
}
\arguments{
\item{cafe.id.line}{the string representing the line number three in the
cafe result file}

\item{species.name}{the species name as given in 'cafe.id.line' to extract
the CAFE internal number for}
}
\value{
The CAFE internal node number representing the species' branch in
the phylogeny. Returned as string.
}
\description{
Extracts the number CAFE assigned the species during its analysis.
}
