% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{getSpliceVariantSeqs}
\alias{getSpliceVariantSeqs}
\title{Identifies those entries in argument \code{seq.lst} that have identical
names as argument \code{seq.names}. If for not all members of
\code{seq.names} entries are found in \code{seq.lst} the function tries to
match those sequence names that were not found with sequence names in
\code{seq.lst} by allowing for splice variants. E.g. \code{GeneA} is not
found in \code{seq.lst}, but \code{GeneA.1} might be.}
\usage{
getSpliceVariantSeqs(seq.lst, seq.names, ...)
}
\arguments{
\item{seq.lst}{An instance of \code{base::list} holding biological sequences
indexed by their unique names (IDs).}

\item{seq.names}{A character vector of sequence IDs to select from argument
\code{seq.lst}.}

\item{...}{Additional arguments to be passed to \code{base::grepl} can be
e.g. \code{ignore.case = TRUE}.}
}
\value{
An instance of \code{base::list} the subset of argument
\code{seq.lst} whose entries match the sequence IDs given in
\code{seq.names}.
}
\description{
Identifies those entries in argument \code{seq.lst} that have identical
names as argument \code{seq.names}. If for not all members of
\code{seq.names} entries are found in \code{seq.lst} the function tries to
match those sequence names that were not found with sequence names in
\code{seq.lst} by allowing for splice variants. E.g. \code{GeneA} is not
found in \code{seq.lst}, but \code{GeneA.1} might be.
}
