% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{hasInvalidVecSpaceComponents.matrix}
\alias{hasInvalidVecSpaceComponents.matrix}
\title{Generic function to check wether any vector component is non-numerical.}
\usage{
\method{hasInvalidVecSpaceComponents}{matrix}(x)
}
\arguments{
\item{x}{A numeric matrix in which each row is interpreted as a single
vector.}
}
\value{
TRUE if and only if any component in any vector in \code{x} is
non-numerical.
}
\description{
Generic function to check wether any vector component is non-numerical.
}
