% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{hasInvalidVecSpaceComponents.numeric}
\alias{hasInvalidVecSpaceComponents.numeric}
\title{Generic function to check wether any vector component is non-numerical.}
\usage{
\method{hasInvalidVecSpaceComponents}{numeric}(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
TRUE if and only if any component in any vector in \code{x} is
non-numerical.
}
\description{
Generic function to check wether any vector component is non-numerical.
}
