% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{hasInvalidVecSpaceComponents}
\alias{hasInvalidVecSpaceComponents}
\title{Generic function to check wether any vector component is non-numerical.}
\usage{
hasInvalidVecSpaceComponents(x)
}
\arguments{
\item{x}{generic object}
}
\value{
TRUE if and only if any component in any vector in \code{x} is
non-numerical.
}
\description{
Generic function to check wether any vector component is non-numerical.
}
