% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{homologs}
\alias{homologs}
\title{Identify homologous gene pairs in a table of significantly similar genes.}
\usage{
homologs(seq.sim.tbl)
}
\arguments{
\item{seq.sim.tbl}{a data frame of two character columns and a third score
column. It is highly important that this table represents only gene pairs of
TWO species, not more or less.}
}
\value{
A subset of the argument data frame, holding only the two character
columns, containing the candidate homologs.
}
\description{
Identify homologous gene pairs in a table of significantly similar genes.
}
