% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\docType{data}
\name{hyphy.meme.bf}
\alias{hyphy.meme.bf}
\title{Brew template to generate a family's MEME input file}
\format{An object of class \code{character} of length 1.}
\usage{
hyphy.meme.bf
}
\description{
Brew template to generate a family's MEME input file
}
\keyword{datasets}
