% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{inferExpectedFalsePositiveRate}
\alias{inferExpectedFalsePositiveRate}
\title{Translates the FUBAR reported expected absolute NUMBER of false positives
into a relative false positive rate for each family within which positively
selected codons have been identified.}
\usage{
inferExpectedFalsePositiveRate(fubar.tbl.res, fubar.expct.fls.pos.n)
}
\arguments{
\item{fubar.tbl.res}{a named vector in which names are families and values
are data.frames of FUBAR results}

\item{fubar.expct.fls.pos.n}{named vector in which names are families and
values are the expected absolute number of false positives as reported by
FUBAR.}
}
\value{
A named vector in which names are families and values are relative
false positive rates per family.
}
\description{
Translates the FUBAR reported expected absolute NUMBER of false positives
into a relative false positive rate for each family within which positively
selected codons have been identified.
}
