% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{interVsIntraSpeciesExpressionProfileDists}
\alias{interVsIntraSpeciesExpressionProfileDists}
\title{Extracts the inter- and intra-species distances from argument object
\code{expr.prof.dists} which is expected to be an instance of class
\code{base::dist} (See \code{expressionProfilesDists} for more details).}
\usage{
interVsIntraSpeciesExpressionProfileDists(expr.prof.dists,
  gene.ids.to.specs.regex = list(ath = "^AT[0-9CM]G\\\\d+$", chi =
  "^CARHR\\\\d+$"))
}
\arguments{
\item{expr.prof.dists}{an instance of \code{base::dist} a single result from
calling \code{expressionProfilesDists}.}

\item{gene.ids.to.specs.regex}{A named character where the names are species
and the values are regular expressions capable of uniquely identifying which
species a gene accession belongs to. Default is \code{list(
'ath'='^AT[0-9CM]G\\d+$', 'chi'='^CARHR\\d+$' )} as within this package's
application we only have expression data for genes of Arabidopsis and C.
hirsuta and in this function's case we are dealing with genes without
expression variant identifiers.}
}
\value{
A list with either data.frame or numeric values. Data.frames have
three columns, two indicating gene pairs and the third the extracted
distances, while numeric values point to maximum distances. Result list
names are the species found as names in \code{gene.ids.to.specs.regex},
\code{inter.species}, \code{max.inter.species}, and for each species
\code{spec_i} one \code{max.spec_i}.
Validate input:
Initialize:
Validate species:
Extract the requested intra-species distances:
Do not extract meaningless distances for pairs of identical
genes, e.g. (gene_a, gene_a):
Extract the requested inter-species distances:
Return result as list:
}
\description{
Extracts the inter- and intra-species distances from argument object
\code{expr.prof.dists} which is expected to be an instance of class
\code{base::dist} (See \code{expressionProfilesDists} for more details).
}
