% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{iprWithSelectedAA}
\alias{iprWithSelectedAA}
\title{Identifies conserved protein domains (InterPro) overapping with the
conserved homologous amino acid (codon) 'sel.aa'.}
\usage{
iprWithSelectedAA(sel.aa, gene, aa.fasta, msa.fasta, iprscan.tbl)
}
\arguments{
\item{sel.aa}{The index of the conserved amino acid under selection in the
corresponding multiple sequence alignment (MSA)}

\item{gene}{The gene accession / ID as used in both 'aa.fasta' and
'msa.fasta'}

\item{aa.fasta}{The result of readAAStringSet(path_2_AAs.fasta) [package
Biostrings]}

\item{msa.fasta}{The result of readAAMultipleAlignment(path_2_AAs_MSA.fasta)
[package Biostrings]}

\item{iprscan.tbl}{The result of calling
readInterProScanResultTable(path_2_interproscan_result_table.tsv)}
}
\value{
A character vector of matching InterPro entries.
}
\description{
Identifies conserved protein domains (InterPro) overapping with the
conserved homologous amino acid (codon) 'sel.aa'.
}

