% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{KaKsStatistics}
\alias{KaKsStatistics}
\title{Uses KaKs_Calculator for selected pairs of aligned coding sequences to infer
Ka and Ks values among other measures. The pairs that are selected are
individuated by \code{GeneFamilies::axtForSelectedDistances}.}
\usage{
KaKsStatistics(cds.msa, gene.group.name, w.dir = tempdir(),
  ka.ks.calculator.path = getOption("GeneFamilies.ka.ks.calculator.path",
  file.path(path.package("GeneFamilies"), "KaKs_Calculator")),
  ka.ks.calculator.model = getOption("GeneFamilies.ka.ks.calculator.model",
  "GMYN"))
}
\arguments{
\item{cds.msa}{The result of invoking \code{seqinr::read.alignment}
representing the multiple alignment of coding sequences.}

\item{gene.group.name}{A string representing the name of the gene group for
which the argument \code{cds.msa} was generated.}

\item{w.dir}{The directory in which to store input and output files with
which the KaKs_Calculator is invoked. Default is \code{tempdir()}.}

\item{ka.ks.calculator.path}{The file path under which to find the
KaKs_Calculator executable. Default is
\code{getOption("GeneFamilies.ka.ks.calculator.path",
file.path(path.package("GeneFamilies"), "KaKs_Calculator"))}.}

\item{ka.ks.calculator.model}{A string used as a command line argument for
the \code{-m} switch of KaKs_Calculator. This argument defines the evolution
model to be used. Default is
\code{getOption("GeneFamilies.ka.ks.calculator.model", "GMYN")}.}
}
\value{
An instance of \code{base::data.frame} the read in tabular result of
invoking KaKs_Calculator using the \code{system} function. There is a
strange behavior of KaKs_Calculator, sometimes it write out a header line to
its result table and sometimes it does not. The result read in here is
assuming NO header line. To get your desired results look for the
appropriate row-name in the first column.
}
\description{
Uses KaKs_Calculator for selected pairs of aligned coding sequences to infer
Ka and Ks values among other measures. The pairs that are selected are
individuated by \code{GeneFamilies::axtForSelectedDistances}.
}
