% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{loadDnaFasta}
\alias{loadDnaFasta}
\title{Simple wrapper around seqinr::read.fasta for DNA data. Function deletes gene
descriptions following the first blank character after the gene accession.
For example the name '>AT1G12345 crazy horse gymnast gene' becomes
'AT1G12345'.}
\usage{
loadDnaFasta(path.2.dna.fasta, sanitize.names = TRUE)
}
\arguments{
\item{path.2.dna.fasta}{The valid file path to the FASTA file to read in.}

\item{sanitize.names}{If TRUE, the default, gene descriptions following the
first blank after the gene accessions are discarded.}
}
\value{
An instance of base::list as returned by \code{seqinr::read.fasta}
}
\description{
Simple wrapper around seqinr::read.fasta for DNA data. Function deletes gene
descriptions following the first blank character after the gene accession.
For example the name '>AT1G12345 crazy horse gymnast gene' becomes
'AT1G12345'.
}
