% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{matchingSpliceVariant}
\alias{matchingSpliceVariant}
\title{Assuming 'gene.ids' holds gene identifiers that lack splice variant suffixes
like '.1' but refer to genes contained in this packages coding sequences
(data) this function aims to identify the matching splice variants.}
\usage{
matchingSpliceVariant(gene.ids, genes = as.character(unlist(spec.gene.ids)))
}
\arguments{
\item{'gene.ids'}{character vector of gene identifiers to be matched to
coding sequence identifiers held in this packages data.}

\item{'genes'}{character vector of gene identifiers with splice variant
suffixes to be matched against the argument gene identifiers.}
}
\value{
Character vector of the original gene identifiers or the matching
splice variants if the original argument identifiers were not found in
'genes'. NA wherever the argument in 'gene.ids' was not found in 'genes' and
could also not be matched to an entry of 'genes'.
}
\description{
Assuming 'gene.ids' holds gene identifiers that lack splice variant suffixes
like '.1' but refer to genes contained in this packages coding sequences
(data) this function aims to identify the matching splice variants.
}
