% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{maxMinusMin}
\alias{maxMinusMin}
\title{Computes the absolute distance between maximum and minimum, optionally
removing \code{NA}.}
\usage{
maxMinusMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric}

\item{na.rm}{If set to \code{TRUE}, the default, \code{NA} values in
\code{x} are removed before computing \code{max} and \code{min}.}
}
\value{
A single numeric value
}
\description{
Computes the absolute distance between maximum and minimum, optionally
removing \code{NA}.
}
