% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{mclDataFrameAsList}
\alias{mclDataFrameAsList}
\title{Converts the data.frame of gene families into a list of gene families.}
\usage{
mclDataFrameAsList(mcl.df)
}
\arguments{
\item{mcl.df}{A data frame with two columns 'Family' and 'Gene' mapping the
families to their gene members. See \code{readMclOutput} for more details.}
}
\value{
A list with names the families in argument \code{mcl.df} and values
the respective gene members joined into sub-lists.
}
\description{
Converts the data.frame of gene families into a list of gene families.
}
