% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\docType{data}
\name{mr.bayes.cds.tree}
\alias{mr.bayes.cds.tree}
\title{Brew template to generate an Mr Bayes input file for a multiple alignment of
Coding Sequences (CDS MSA).}
\format{An object of class \code{character} of length 1.}
\usage{
mr.bayes.cds.tree
}
\description{
Brew template to generate an Mr Bayes input file for a multiple alignment of
Coding Sequences (CDS MSA).
}
\keyword{datasets}
