% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\docType{data}
\name{mr.bayes.mixed.prot.tree}
\alias{mr.bayes.mixed.prot.tree}
\title{Brew template to generate a protein group's Mr Bayes input file using
additional e.g. morphological knowledge}
\format{An object of class \code{character} of length 1.}
\usage{
mr.bayes.mixed.prot.tree
}
\description{
Brew template to generate a protein group's Mr Bayes input file using
additional e.g. morphological knowledge
}
\keyword{datasets}
