% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\docType{data}
\name{mr.bayes.prot.tree}
\alias{mr.bayes.prot.tree}
\title{Brew template to generate a protein group's Mr Bayes input file}
\format{An object of class \code{character} of length 1.}
\usage{
mr.bayes.prot.tree
}
\description{
Brew template to generate a protein group's Mr Bayes input file
}
\keyword{datasets}
