% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{mrBayesTreeFromMixedMSA}
\alias{mrBayesTreeFromMixedMSA}
\title{Generates a valid input file for Mr Bayes, which will generate a Bayesian
phylogeny from a mixed multiple sequence alignment, e.g. CDS AND an alignment of
additional e.g. morphological data.}
\usage{
mrBayesTreeFromMixedMSA(msa.lst, msa.types, mr.bayes.dir, gene.group.name,
  mr.bayes.n.chains = getOption("mc.cores", 4))
}
\arguments{
\item{msa.lst}{A list of multiple sequence alignments, a mix of DNA and
Standard data}

\item{msa.types}{a character vector of the types of each MSA: DNA, Codon, or
Standard}

\item{mr.bayes.dir}{The file directory in which to store the output file}

\item{the}{name of the analysis. Will be the head of the file name holding
the input for Mr Bayes}
}
\description{
Generates a valid input file for Mr Bayes, which will generate a Bayesian
phylogeny from a mixed multiple sequence alignment, e.g. CDS AND an alignment of
additional e.g. morphological data.
}
