% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{naAsZero}
\alias{naAsZero}
\title{Helper function that converts argument \code{x} to the number zero if and
only if \code{is.na(x)}.}
\usage{
naAsZero(x)
}
\arguments{
\item{x}{any scalar}
}
\value{
\code{0.0} if \code{is.na(x)} or \code{x} otherwise
}
\description{
Helper function that converts argument \code{x} to the number zero if and
only if \code{is.na(x)}.
}
