% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{nonConservedCompositeAnnotations}
\alias{nonConservedCompositeAnnotations}
\title{Identifies those composite annotations that are less frequent than the most
frequent ones found for the argument genes and based upon argument gene
(function) annotations. For more details about composite annotations see
function 'compositeAnnotations(...)'}
\usage{
nonConservedCompositeAnnotations(gene.accs, gene.annos = all.ipr,
  gene.col = 1, anno.col = 2)
}
\arguments{
\item{gene.accs}{The identifiers or accessions of the genes to compute the
entropy for}

\item{gene.annos}{The data.frame holding the annotations for the genes in
'gene.accs'. Default is all available InterPro annotations expected to be
found in 'all.ipr'}

\item{gene.col}{the column of 'gene.annos' in which to lookup the gene
identifiers or gene accessions. Default is 1}

\item{anno.col}{the column of 'gene.annos' in which to lookup the function
annotation for the genes in 'gene.accs'. Default is 2}
}
\value{
An instance of base::table
}
\description{
Identifies those composite annotations that are less frequent than the most
frequent ones found for the argument genes and based upon argument gene
(function) annotations. For more details about composite annotations see
function 'compositeAnnotations(...)'
}
