% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{normalizeVector}
\alias{normalizeVector}
\title{Normalizes the vector \code{x} to length 1.0.}
\usage{
normalizeVector(x)
}
\arguments{
\item{numeric}{vector}
}
\value{
\code{x / euclNorm(x)} or \code{NaN} if
\code{hasInvalidVecSpaceComponents(x)} is \code{TRUE}.
}
\description{
Normalizes the vector \code{x} to length 1.0.
}
