% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{numberAnnotations}
\alias{numberAnnotations}
\title{Counts the number of times an annotation, e.g. a Pfam or InterPro entry, is
annotated in argument annotation table 'anno.tbl'.}
\usage{
numberAnnotations(anno, anno.tbl)
}
\arguments{
\item{anno}{string contained in column 2 of 'anno.tbl'}

\item{anno.tbl}{a data.frame of two columns, column 1 holds gene identifiers
and column 2 annotations, e.g. InterPro or Pfam entries.}
}
\value{
integer - the number of genes 'anno' has been found annotated to.
}
\description{
Counts the number of times an annotation, e.g. a Pfam or InterPro entry, is
annotated in argument annotation table 'anno.tbl'.
}
