% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{orthologs}
\alias{orthologs}
\title{Identifies orthologous gene pairs as genes being significantly similar in
sequence and belonging to different species.}
\usage{
orthologs(homologs.tbl, gene.regex.1, gene.regex.2)
}
\arguments{
\item{homologs.tbl}{a data.frame of two character columns - possibly the
result obtained from invoking homologous(...)}

\item{gene.regex.1}{a string encoding a PERL regular expression to match
gene identifiers to species A}

\item{gene.regex.2}{a string encoding a PERL regular expression to match
gene identifiers to species B}
}
\value{
A subset of argument 'homologs.tbl' containg the orthologous gene
pairs. Note that if pair (A,B) is contained the reciprocal pair (B,A) is,
too. This should ease lookup.
}
\description{
Identifies orthologous gene pairs as genes being significantly similar in
sequence and belonging to different species.
}
