% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{orthologsFromBLATpslTable}
\alias{orthologsFromBLATpslTable}
\title{Parses 'psl' default BLAT output formatted files to infer orthologs between
species. Uses PERL regular expressions to identify which gene identifiers
belongs to which species. Hence by using two such regular expressions
orthologs between two species can be extracted, even though the BLAT output
file contains similarity pairs from more than two species. Note that
orthology is inferred based on reciprocal best search results.}
\usage{
orthologsFromBLATpslTable(blat.df, query.col = 10, target.col = 14,
  matching.bases.col = 1, qSize.col = 11, tSize.col = 15,
  spec.regexs = c("CAHR\\\\d+(\\\\.\\\\d)?",
  "AT[0-9MC]G\\\\d+(\\\\.\\\\d)?"))
}
\arguments{
\item{blat.df}{a data frame holding the output of BLAT in psl format. Se
function readBlatPSLoutput(...) for more details}

\item{query.col}{the number of the column in which to lookup the query
sequence's name}

\item{target.col}{the number of the column in which to lookup the target
sequence's name}

\item{matching.bases.col}{the number of the column in which to lookup the
number of nucleotide bases matching between query and target}

\item{qSize.col}{the number of the column in which to lookup the query
sequence's size}

\item{tSize.col}{the number of the column in which to lookup the target
sequence's size}

\item{spec.regexs}{a character vector containg two regular expressions
matching gene identifiers from each of the two species orthologies are to be
found for}
}
\value{
A data frame of two columns holding the gene identifiers of the
orthologous gene pairs. Note, that for reasons of easing lookup inverse
pairs are also contained.
}
\description{
Parses 'psl' default BLAT output formatted files to infer orthologs between
species. Uses PERL regular expressions to identify which gene identifiers
belongs to which species. Hence by using two such regular expressions
orthologs between two species can be extracted, even though the BLAT output
file contains similarity pairs from more than two species. Note that
orthology is inferred based on reciprocal best search results.
}
