% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{orthologsFromPairwiseReciprocalBestHits}
\alias{orthologsFromPairwiseReciprocalBestHits}
\title{Extracts clusters of highly conserved orthologous genes from pairwise
reciprocal best hits obtained from sequence similarity searches. The
searches are explained in detail in this package's Vignette
\code{GeneFamilies}, section \code{Ortholog Identification}. Only ortholog
clusters that have the exact same size as the numer of investigated species
are returned.}
\usage{
orthologsFromPairwiseReciprocalBestHits(pair.best.hits.tbl,
  spec.gene.ids.arg = spec.gene.ids, anchor.gene.col = 1,
  candidate.gene.col = 2)
}
\arguments{
\item{pair.best.hits.tbl}{An instance of \code{base::data.frame} holding the
final result of the concatonated filtered pairwise sequence similarity
searches. See Vignette \code{GeneFamilies} for details on how to generate
this input table.}

\item{spec.gene.ids.arg}{An instance of \code{base::list} with names being
the investigated species and values all protein accessions belonging to the
respective species. Default is \code{spec.gene.ids} (see file \code{zzz.R}
for more details).}

\item{anchor.gene.col}{The column-name or column-index of
\code{pair.best.hits.tbl} in which to find the anchor genes, i.e. those
genes of the species all pairwise sequence similarity searches were
conducted with. Default is \code{1}.}

\item{candidate.gene.col}{The column-name or column-index of
\code{pair.best.hits.tbl} in which to find the candidate genes, that were
the best hits in the pairwise sequence similarity searches. Default is
\code{2}.}
}
\value{
An instance of \code{base::data.frame} with one ortholog cluster per
row and one column for each of the investigated species. An additional
column holds the ortholog cluster name, composed as
\code{'ortholog_cluster_<number>'}.
}
\description{
Extracts clusters of highly conserved orthologous genes from pairwise
reciprocal best hits obtained from sequence similarity searches. The
searches are explained in detail in this package's Vignette
\code{GeneFamilies}, section \code{Ortholog Identification}. Only ortholog
clusters that have the exact same size as the numer of investigated species
are returned.
}
