% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{parseCafeBranchSpecificPValues}
\alias{parseCafeBranchSpecificPValues}
\title{Extract the branch specific extension / contraction P-Values that were
assigned to those branches that lead to the species' tips in the phylogeny.}
\usage{
parseCafeBranchSpecificPValues(cafe.br.spec.pvals, spec.pos)
}
\arguments{
\item{cafe.br.spec.pvals}{the string parsed out from the CAFE result table}

\item{spec.pos}{named integer vector holding the positions of the species
specific p-values in 'cafe.br.spec.pvals'. See ./exec/parseCafeResult.R for
more details}
}
\value{
A data frame with columns the species and cells their specific
P-Values for expansion or contraction in the respective gene family.
Do not warn about the intent to make a numeric NA from the CAFE '-'
character:
}
\description{
Extract the branch specific extension / contraction P-Values that were
assigned to those branches that lead to the species' tips in the phylogeny.
}
