% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{parseMemeSiteResult}
\alias{parseMemeSiteResult}
\title{Reads in the tabular MEME site result table and adds a column with adjusted
P-Values.}
\usage{
parseMemeSiteResult(meme.out, p.adjust.method = "BY")
}
\arguments{
\item{meme.out}{A valid file-path to the MEME tabular site output file}

\item{p.adjust.method}{The argument provided to \code{base::p.adjust} as the
\code{method} used to adjust the p-values. Default is \code{"BY"}.}
}
\value{
An instance of data.frame read from the MEME site tabular output
with an added \code{Site} and adjusted p-values (\code{p.adj}) columns.
}
\description{
Reads in the tabular MEME site result table and adds a column with adjusted
P-Values.
}

