% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{passSupportValues}
\alias{passSupportValues}
\title{Passes support values from inner nodes in argument 'trees' to matching inner
nodes in argument 'tree', if and only if the respective inner nodes span a
subtree terminating in identical sets of tips. Note, that the inner topology
is ignored.}
\usage{
passSupportValues(tree, trees, nds = unique(tree$edge[, 1]))
}
\arguments{
\item{tree}{an instance of class ape::phylo}

\item{trees}{a vector of instances of class ape::phylo}

\item{nds}{an integer vector indicating inner nodes of the argument 'tree'
to process}
}
\value{
An annotated copy of argument 'tree'
}
\description{
Passes support values from inner nodes in argument 'trees' to matching inner
nodes in argument 'tree', if and only if the respective inner nodes span a
subtree terminating in identical sets of tips. Note, that the inner topology
is ignored.
}
