% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{perpVecToDiagonal}
\alias{perpVecToDiagonal}
\title{Drops a perpendicular from vector \code{x} to the vector space's diagonal.
The perpendicular points towards \code{x}.}
\usage{
perpVecToDiagonal(x, d.v = rep(1, length(x)))
}
\arguments{
\item{x}{numeric representing a vector in n-dimensional space}

\item{d.v}{numeric the diagonal vector in n-dimensional space. Default is
\code{rep(1, length(x))}.}
}
\value{
A vector in the same space of \code{x}, the perpendicular.
}
\description{
Drops a perpendicular from vector \code{x} to the vector space's diagonal.
The perpendicular points towards \code{x}.
}
