% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpeciesSpecificAnnotationDiversity}
\alias{plotSpeciesSpecificAnnotationDiversity}
\title{Scatter-Plots the pecies specific annotation (function) diversity for each
single gene group found in 'entropy.df'.}
\usage{
plotSpeciesSpecificAnnotationDiversity(plot.pdf, entropy.df, species)
}
\arguments{
\item{plot.pdf}{The PDF File into which to save the scatter-plot}

\item{entropy.df}{A data.frame with species specific annotation entropies
per gene group. Gene groups are rows and columns need to include the species
in argument 'species'.}

\item{species}{A character of length 2 holding the names of the two species
to use for the plot. The species names need to occur as column names in
'entropy.df'.}
}
\value{
TRUE if and only if the plot can be generated successfully
}
\description{
Scatter-Plots the pecies specific annotation (function) diversity for each
single gene group found in 'entropy.df'.
}
