% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpeciesSpecificAnnotationDiversityDists}
\alias{plotSpeciesSpecificAnnotationDiversityDists}
\title{Boxplots the distributions of species specific annotation (function)
diversity, measured for each single gene group found in 'entropy.df'.}
\usage{
plotSpeciesSpecificAnnotationDiversityDists(plot.pdf, entropy.df, species,
  gene.groups.type, star.species = c(),
  colors = brewer.pal(length(species), "Dark2"))
}
\arguments{
\item{plot.pdf}{The PDF File into which to save the distributions}

\item{entropy.df}{A data.frame with species specific annotation entropies
per gene group. Gene groups are rows and columns need to include the species
in argument 'species'.}

\item{species}{A character of length 2 or greater holding the names of the species
to use for the plot. The species names need to occur as column names in
'entropy.df'.}

\item{gene.groups.type}{The type of gene groups: e.g. 'Tandem Clusters'.
Will be used in the main title of the boxplot.}

\item{star.species}{The names of those species above whose boxplots to print
a '*'. Default is none (\code{c()}).}

\item{colors}{A character vector holding the colors for each species'
density. Default is the brewer palette 'Dark2' for length(species).}
}
\value{
TRUE if and only if the plot can be generated successfully
}
\description{
Boxplots the distributions of species specific annotation (function)
diversity, measured for each single gene group found in 'entropy.df'.
}
