% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{plotTissueSpecificExpressionProfileDistanceDistribs}
\alias{plotTissueSpecificExpressionProfileDistanceDistribs}
\title{Plots horizontal boxplots of per tissue distributions of expression profile
distances. Draws a line from each upper Whisker to the next in order to
highlight the hourglass shape. Tissues must be given in the order according
to a plant's life.}
\usage{
plotTissueSpecificExpressionProfileDistanceDistribs(plot.df, tissues,
  gene.classes, pdf.path, plot.title = "")
}
\arguments{
\item{plot.df}{A data.frame with the following columns: \code{'Tissue',
'Gene.Class', 'Value'}.}

\item{tissues}{A character vector of tissues held in \code{plot.df$Tissue}.
The order found in this argument will be the order on the Y-Axis of the
horizontal plot.}

\item{gene.classes}{A character vector of gene.classes to group paired
boxplots by. Must be of length 2.}

\item{pdf.path}{The valid file path to the PDF in which the horizontal
boxplot is to be stored.}

\item{plot.title}{A string defining the headline printedabove the plot. See
argument \code{main} in function \code{base::boxplot} for more details.
Default is \code{''}.}
}
\value{
The output of \code{dev.off()}
}
\description{
Plots horizontal boxplots of per tissue distributions of expression profile
distances. Draws a line from each upper Whisker to the next in order to
highlight the hourglass shape. Tissues must be given in the order according
to a plant's life.
}
