% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseKaKs.R
\name{putGenePairIntoRedis}
\alias{putGenePairIntoRedis}
\title{Puts the integer value 1 into the Redis Server using the alphabetically
sorted gene accessions in 'gene.pair' as key.}
\usage{
putGenePairIntoRedis(gene.pair)
}
\arguments{
\item{Character}{vector of length two holding members A and B of the gene
pair.}
}
\value{
The output of calling rredis::redisSet(...)
}
\description{
Puts the integer value 1 into the Redis Server using the alphabetically
sorted gene accessions in 'gene.pair' as key.
}
