% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{rad2deg}
\alias{rad2deg}
\title{Converts an angle's size given in \code{radians} to \code{degrees}.}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{a numeric}
}
\value{
The converted \code{rad} value in degrees.
}
\description{
Converts an angle's size given in \code{radians} to \code{degrees}.
}
