% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{readBlatPSLoutput}
\alias{readBlatPSLoutput}
\title{Wraps function base::read.table with the proper arguments needed to
efficiently and correctly read in a BLAT output table in PSL format.}
\usage{
readBlatPSLoutput(blat.psl.path)
}
\arguments{
\item{blat.psl.path}{valid file path to the respective BLAT psl output file}
}
\value{
The read in table as data frame
}
\description{
Wraps function base::read.table with the proper arguments needed to
efficiently and correctly read in a BLAT output table in PSL format.
}
