% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{readFubarTable}
\alias{readFubarTable}
\title{Parses a FUBAR output table for significant posterior probabilities of a
codon being subject to positive selection. If any codons fit the criteria, a
respective table is returned, otherwise \code{NULL} is returned.}
\usage{
readFubarTable(path.2.fubar.csv, family.name, sign.post.prob = 0.9)
}
\arguments{
\item{path.2.fubar.csv}{a valid file path to the FUBAR csv file to parse}

\item{family.name}{A string indicating the name of the gene cluster (family)
subjected to the FUBAR analysis.}

\item{sign.post.prob}{The cutoff value the posterior probabilities have to
meet (\code{>=}) in order to be considered significant. Default is
\code{.9}.}
}
\value{
An instance of base::data.frame holding the tabular FUBAR output
with just the following columns: 1. Codon, 2. Post.Prob, 3. BayesFactor, and
4. family. \code{NULL} is returned if no codon matches the significance
criteria.
}
\description{
Parses a FUBAR output table for significant posterior probabilities of a
codon being subject to positive selection. If any codons fit the criteria, a
respective table is returned, otherwise \code{NULL} is returned.
}
