% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{readInterProScanResultTable}
\alias{readInterProScanResultTable}
\title{Parses the result file of InterProScan into a data.frame.}
\usage{
readInterProScanResultTable(path.2.iprscan.res)
}
\arguments{
\item{path.2.iprscan.res}{The file path to the argument InterProScan result
table.}
}
\value{
A data.frame with 14 columns.
}
\description{
Parses the result file of InterProScan into a data.frame.
}
