% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{readKaKsCalculatorOutput}
\alias{readKaKsCalculatorOutput}
\title{Uses base::read.table to parse the output of 'KaKs_Calculator'.}
\usage{
readKaKsCalculatorOutput(path.2.KaKs.Calc.output)
}
\arguments{
\item{path.2.KaKs.Calc.output}{the valid file path to the output table to be
parsed}
}
\value{
An instance of base::data.frame with three columns: 'Sequence',
'Ka.Ks', 'P.Value.Fisher.'
}
\description{
Uses base::read.table to parse the output of 'KaKs_Calculator'.
}
