% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{readMclOutput}
\alias{readMclOutput}
\title{Uses Linux's AWK and package data.table to parse the output of MCL into an
instance of base::data.frame.}
\usage{
readMclOutput(path.2.mcl.out, family.name.prefix = "cluster_")
}
\arguments{
\item{path.2.mcl.out}{The valid path to the output file of the \code{mcl}
program.}

\item{family.name.prefix}{The prefix used to name gene families (clusters).
Default is \code{'cluster_'} resulting in 'cluster_1', 'cluster_2', etc.}
}
\value{
A data.frame with two columns: 'Family', and 'Gene' mapping the
families to their gene members.
}
\description{
Uses Linux's AWK and package data.table to parse the output of MCL into an
instance of base::data.frame.
}
