% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{readMinimumInterProScanResultTable}
\alias{readMinimumInterProScanResultTable}
\title{Uses function 'readInterProScanResultTable(path.2.iprscan.res)' and retains
only two columns: 1. Gene Identifier, and 2. InterPro Identifier.}
\usage{
readMinimumInterProScanResultTable(path.2.iprscan.res, gene.col = 1,
  ipr.col = 12)
}
\arguments{
\item{path.2.iprscan.res}{The file path to the argument InterProScan result
table.}

\item{gene.col}{The column identifier for the column in which to find the
gene identifiers.}

\item{ipr.col}{The column identifier for the column in which to find the
InterPro identifiers.}
}
\value{
An instance of base::data.frame
}
\description{
Uses function 'readInterProScanResultTable(path.2.iprscan.res)' and retains
only two columns: 1. Gene Identifier, and 2. InterPro Identifier.
}
