% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{removeDuplicatedInversePairs}
\alias{removeDuplicatedInversePairs}
\title{Identifies duplicated inverse pairs and discards them.}
\usage{
removeDuplicatedInversePairs(clst.hmlgs.list, ret.as.list = TRUE)
}
\arguments{
\item{clst.hmlgs.list}{a named list in which for each pair member A is the
name and member B is the value}

\item{ret.as.list}{indicates whether to return the result as a list, subset
of the original 'clst.hmlgs.list', or as a character matrix of two columns one
column per member. For the latter option set 'ret.as.list' to FALSE.}
}
\value{
The subset of pairs that are not inverse duplicates, either as list
or as character matrix.
}
\description{
Identifies duplicated inverse pairs and discards them.
}
