% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{removeExpressionVariant}
\alias{removeExpressionVariant}
\title{Modifies the argument \code{gene.ids} by deleting expression variant
identifiers \emph{and} removing resulting duplicate gene IDs. See
\code{base::sub} for more details.}
\usage{
removeExpressionVariant(gene.ids,
  reg.ex = getOption("MaizeGeneFamilies.expression.variant.regex",
  "\\\\.\\\\d+$"))
}
\arguments{
\item{gene.ids}{a character vector of gene identifiers.}

\item{reg.ex}{the regular expression used to identify the to be deleted
expression variant part of a gene ID. Default is
\code{getOption('MaizeGeneFamilies.expression.variant.regex', '\\.\\d+$')}.}
}
\value{
The modified gene IDs with the expression variant parts deleted.
}
\description{
Modifies the argument \code{gene.ids} by deleting expression variant
identifiers \emph{and} removing resulting duplicate gene IDs. See
\code{base::sub} for more details.
}

