% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{removeNodeLabelsAndBranchLengths}
\alias{removeNodeLabelsAndBranchLengths}
\title{Removes node labels and branch lengths from instance of ape::phylo}
\usage{
removeNodeLabelsAndBranchLengths(phyl.tree)
}
\arguments{
\item{phyl.tree}{instance of class ape::phylo}
}
\value{
A copy of 'phyl.tree' without branch length and without node labels
}
\description{
Removes node labels and branch lengths from instance of ape::phylo
}
