% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{removeNonAAChars}
\alias{removeNonAAChars}
\title{Removes all non Amino Acid characters from an aligned amino acid sequence.}
\usage{
removeNonAAChars(aa.seq, non.aa.regex = "[-]")
}
\arguments{
\item{aa.seq}{The aligned amino acid sequence, possibly including to be
removed characters.}

\item{non.aa.regex}{A PERL regular expression used to identify to be removed
matches from 'aa.seq'}
}
\value{
A string in which all matches have been deleted.
}
\description{
Args:
}
