% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{removeSpliceVariant}
\alias{removeSpliceVariant}
\title{Removes trailing splice variant identifier from gene IDs; e.g. CARHR213450.1
will result in CARHR213450}
\usage{
removeSpliceVariant(gene.ids)
}
\arguments{
\item{gene.ids}{a character vector of gene identifiers}
}
\value{
The modified 'gene.ids' without the trailing splice variant IDs.
}
\description{
Removes trailing splice variant identifier from gene IDs; e.g. CARHR213450.1
will result in CARHR213450
}
