% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseKaKs.R
\name{renderCodemlInputs}
\alias{renderCodemlInputs}
\title{Generates the tree and control file to run PAML's codeml in order to compute
the Ks of 'gene.pair'}
\usage{
renderCodemlInputs(gene.pair, cds.msa.path,
  ks.out.path = sub("_CDS_MSA.fasta", "_codeml_out.txt", cds.msa.path,
  fixed = TRUE), cds.codeml.path = sub("_CDS_MSA.fasta",
  "_codeml_in.cnt", cds.msa.path, fixed = TRUE),
  tree.path = sub("_CDS_MSA.fasta", "_tree.newick", cds.msa.path, fixed =
  TRUE))
}
\arguments{
\item{gene.pair}{a character of length 2 holding the identifiers of the
pair's genes.}

\item{cds.msa.path}{the file path to the gene pair's codon sequences
alignment}

\item{ks.out.path}{the file path to the gene pair's codeml output file,
defaut is generated from 'cds.msa.path' substituting '_CDS_MSA.fasta' with
'_codeml_out.txt'}

\item{cds.codeml.path}{the file path to the gene pair's codeml input file,
defaut is generated from 'cds.msa.path' substituting '_CDS_MSA.fasta' with
'_codeml_in.cnt'}

\item{tree.path}{the file path to the gene pair's tree file, defaut is
generated from 'cds.msa.path' substituting '_CDS_MSA.fasta' with
'_tree.newick'. The tree will be the trivial one: '(A,B);'.}
}
\value{
A list with two string entries: 'in' the path to the control file
used as input to PAML's codeml, and 'out' path to the output file generated
by PAML's codeml.
}
\description{
Generates the tree and control file to run PAML's codeml in order to compute
the Ks of 'gene.pair'
}
