% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{replaceSanitizedWithOriginalIDs}
\alias{replaceSanitizedWithOriginalIDs}
\title{Replaces sanitized protein identifiers in XStringSet 'xstring.set' with the
original ones held in 'name.maps' table.}
\usage{
replaceSanitizedWithOriginalIDs(xstring.set, name.maps,
  san.col = "sanitized", orig.col = "original")
}
\arguments{
\item{xstring.set}{A vector or list of strings}

\item{name.maps}{A data.frame with at least two columns one holding the
sanitized and the other the original protein IDs}

\item{san.col}{The column index or name of 'name.maps' holding the sanitized
protein IDs}

\item{orig.col}{The column index or name of 'name.maps' holding the original
protein IDs}
}
\value{
A copy of argument 'xstring.set' with the sanitized protein IDs
replaced with their originals.
}
\description{
Replaces sanitized protein identifiers in XStringSet 'xstring.set' with the
original ones held in 'name.maps' table.
}
