% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{replaceWithOriginal}
\alias{replaceWithOriginal}
\title{Returns the matching original names for the sanitized ones held in argument
'san.names'.}
\usage{
replaceWithOriginal(san.names, name.mappings)
}
\arguments{
\item{'san.names'}{a character vector of sanitized names (gene accessions)}

\item{'name.mappings'}{a data.frame generated by function
'sanitizeNames(...)'. This data.frame is used to lookup the matching
original names for the argument sanitized names.}
}
\value{
A character vector in which the original names are stored in the
order of their matching sanitized ones as they appear in the argument
'san.names'.
}
\description{
Returns the matching original names for the sanitized ones held in argument
'san.names'.
}
