% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{replaceWithOriginalInFigTree}
\alias{replaceWithOriginalInFigTree}
\title{Uses GNU sed to replace the sanitized gene identifiers with the original
ones in the Fig-Tree file - consensus tree generated by Mr. Bayes.}
\usage{
replaceWithOriginalInFigTree(name.mappings, path.2.fig.tree,
  path.2.fig.tree.orig = sub(".con.tre", "_original_identifiers.con.tre",
  path.2.fig.tree, fixed = TRUE))
}
\arguments{
\item{name.mappings}{data.frame with two columns 'original', and
'sanitized', holding the pairs of original gene ID to sanitized gene ID.}

\item{path.2.fig.tree}{the valid file path to the tree file to process}

\item{path.2.fig.tree.orig}{the valid file path to the output file in which
the original gene identifiers are used. Must not be identical to
'path.2.fig.tree'. Default is 'path.2.fig.tree' with its suffix '.con.tre'
modified to '_original_identifiers.con.tre'.}
}
\description{
Uses GNU sed to replace the sanitized gene identifiers with the original
ones in the Fig-Tree file - consensus tree generated by Mr. Bayes.
}
