% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Many programs in Bioinformatics cannot handle gene accessions with special
characters. That is why we rename the genes in a list as generated by
\code{seqinr::read.fasta} with 'PROT1', 'PROT2'...}
\usage{
sanitizeNames(xstring.set)
}
\arguments{
\item{'xstring.set'}{an instance of base::list as generated by
\code{seqinr::read.fasta}}
}
\value{
An instance of data.frame with two columns the 'original' and the
'sanitized' names.
}
\description{
Many programs in Bioinformatics cannot handle gene accessions with special
characters. That is why we rename the genes in a list as generated by
\code{seqinr::read.fasta} with 'PROT1', 'PROT2'...
}
