% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{scalarProjection}
\alias{scalarProjection}
\title{Computes the scalar projection of vector \code{a} onto vector \code{b}, i.e.
the length of the segment of \code{b} 'overshadowed' by \code{a}. See
en.wikipedia.org/wiki/Vector_projection}
\usage{
scalarProjection(a, b)
}
\arguments{
\item{a}{numerical representing first vector}

\item{b}{numerical representing second vector}
}
\value{
A scalar value, the length of the segment of \code{b} overshadowed
by \code{a}.
}
\description{
Computes the scalar projection of vector \code{a} onto vector \code{b}, i.e.
the length of the segment of \code{b} 'overshadowed' by \code{a}. See
en.wikipedia.org/wiki/Vector_projection
}
