% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{selectGeneGroupSpecificEnrichedAnnotations}
\alias{selectGeneGroupSpecificEnrichedAnnotations}
\title{Infer enriched annotations specific to the argument type of gene-groups, add
human readable information to it and return it as a data.frame.}
\usage{
selectGeneGroupSpecificEnrichedAnnotations(gene.group.type,
  gene.group.type.enriched.annos = list(positively.selected =
  fams.pos.sel.non.conserved.iprs.enrich.df$Annotation, expanded.contracted
  = exp.fams.df.non.conserved.iprs.enrich.df$Annotation, tandems =
  bra8.tandem.clusters.non.conserved.iprs.enrich.df$Annotation, conserved =
  fams.conserved.non.conserved.iprs.enrich.df$Annotation),
  anno.db = ipr.db, anno.name = "NAME")
}
\arguments{
\item{gene.group.type}{The type of the gene.groups for which to identify
specific enrichments.}

\item{gene.group.type.enriched.annos}{A list of enriched annotations per
gene group. Names must include \code{gene.group.type}. Default is the
enriched annotations found in this packages data.}

\item{anno.db}{A list with names being the annotations as in the names of
\code{enriched.annotations} and values are lists with a slot holding human
readable information about the annotation.}

\item{anno.name}{The slot name of the lists held in \code{anno.db} to
extract for human readable information about the annotations.}
}
\value{
A data.frame with two columns: 1. Annotation and 2. Annotation.Name
}
\description{
Infer enriched annotations specific to the argument type of gene-groups, add
human readable information to it and return it as a data.frame.
}
