% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{selectSignificantEnrichments}
\alias{selectSignificantEnrichments}
\title{Filters 'enrichment.tbl' for those annotations that have a significant
adjusted p.value.}
\usage{
selectSignificantEnrichments(enrichment.tbl, significance.level = 0.05)
}
\arguments{
\item{enrichment.tbl}{a data.frame returned by function 'enrichmentTests'}

\item{significance.level}{the p.value cutoff to apply, default is '<= 0.05'}
}
\value{
A subset of 'enrichment.tbl' holding only the significant results.
}
\description{
Filters 'enrichment.tbl' for those annotations that have a significant
adjusted p.value.
}
