% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{selectSignificantlyEnrichedAnnotations}
\alias{selectSignificantlyEnrichedAnnotations}
\title{Selects those results obtained from calling \code{enrichedAnnotations} that
are significant and adds informative information to the annotations
extracted from the \code{anno.db}.}
\usage{
selectSignificantlyEnrichedAnnotations(enriched.annotations,
  p.val.cutoff = 0.01, anno.db = ipr.db, anno.name = "NAME")
}
\arguments{
\item{enriched.annotations}{A numeric vector with values the P-Values and
names the annotations. See \code{enrichedAnnotations(...)} for details.}

\item{p.val.cutoff}{The significance cuto-off to be applied. Default is 0.01}

\item{anno.db}{A list with names being the annotations as in the names of
\code{enriched.annotations} and values are lists with a slot holding human
readable information about the annotation.}

\item{anno.name}{The slot name of the lists held in \code{anno.db} to
extract for human readable information about the annotations.}
}
\value{
A data.frame with three columns: 1. Annotation, 2. P.Value.Adjusted,
and 3. Annotation.Name
}
\description{
Selects those results obtained from calling \code{enrichedAnnotations} that
are significant and adds informative information to the annotations
extracted from the \code{anno.db}.
}
