% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{spec.regex.gene.ids}
\alias{spec.regex.gene.ids}
\title{####################
Define constants: #
####################
'spec.regex.gene.ids' - Regular expressions to identify the species a gene
belongs to. They are supposed to be used with the perl=TRUE flag.}
\format{An object of class \code{list} of length 8.}
\usage{
spec.regex.gene.ids
}
\description{
####################
Define constants: #
####################
'spec.regex.gene.ids' - Regular expressions to identify the species a gene
belongs to. They are supposed to be used with the perl=TRUE flag.
}
\keyword{datasets}
