% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{speciesComposition}
\alias{speciesComposition}
\title{Infers the number of genes belonging to the given species.}
\usage{
speciesComposition(fam, spec.cds.ids = spec.gene.ids)
}
\arguments{
\item{'fam'}{a character vector of gene identifiers representing the
argument gene family.}

\item{'spec.cds.ids'}{a named list of character vectors indicating which
gene identifiers belong to which species. Default is set to constant
'spec.gene.ids' declared in R/zzz.R}
}
\value{
An integer vector with names equeal to the names of argument list
'spec.cds.ids' and values indicating how many gene members were found in the
respective species.
}
\description{
Infers the number of genes belonging to the given species.
}
