% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_funks.R
\name{speciesCompositionsToDataFrame}
\alias{speciesCompositionsToDataFrame}
\title{Generate a data frame of the list of species composition integer vectors.
Rows are named as families and columns as species.}
\usage{
speciesCompositionsToDataFrame(spec.comps.lst,
  column.names = names(spec.comps.lst[[1]]),
  family.names = names(spec.comps.lst))
}
\arguments{
\item{spec.comps.lst}{- A list of integer vectors generated by lapply
function 'speciesComposition' on the gene families character vectors.}

\item{column.names}{- A character vector of column names. Default is the
names slot of the first entry of 'spec.comps.lst'}

\item{family.names}{- A character vector of family names. Default is the
names slot of argument list 'spec.comps.lst'}
}
\value{
A data frame: One row per family and one column per species with an
extra column of the family sizes.
}
\description{
Generate a data frame of the list of species composition integer vectors.
Rows are named as families and columns as species.
}
