% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{speciesForGeneId}
\alias{speciesForGeneId}
\title{Identifies the species of a gene identifier using the original gene
identifiers in this project's data sets. Note that the constant
spec.gene.ids is used by default to map gene identifiers to species. See
zzz.R for more details.}
\usage{
speciesForGeneId(gene.id, spec.genes = spec.gene.ids)
}
\arguments{
\item{gene.id}{}

\item{spec.genes}{A names list, in which the names indicate the species and
the values are character vectors of gene identifiers.}
}
\value{
the matching species as a name in argument list 'spec.regexs'
}
\description{
Identifies the species of a gene identifier using the original gene
identifiers in this project's data sets. Note that the constant
spec.gene.ids is used by default to map gene identifiers to species. See
zzz.R for more details.
}
