% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{speciesForGeneId_Regex}
\alias{speciesForGeneId_Regex}
\title{Identifies the species of a gene identifier using regular expressions. Note
that the constant spec.regex.gene.ids is used by default to map regular
expression to species. See zzz.R for more details.}
\usage{
speciesForGeneId_Regex(gene.id, spec.regexs = spec.regex.gene.ids)
}
\arguments{
\item{gene.id}{}

\item{spec.regexs}{a list with names indicating the species and values the
regular expressions to match gene identifiers against.}
}
\value{
the matching species as a name in argument list 'spec.regexs'
}
\description{
Identifies the species of a gene identifier using regular expressions. Note
that the constant spec.regex.gene.ids is used by default to map regular
expression to species. See zzz.R for more details.
}
