% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{statVectorCloud}
\alias{statVectorCloud}
\title{Computes the statistic \code{stat} (default is \code{base::mean}) vector of
the cloud of vectors. Also infers the deviance in terms of
\code{deviance.funk} (default is \code{base::sd}) of each axis of the vector
space.}
\usage{
statVectorCloud(vecs.df,
  stat = getOption("GeneFamilies.vector.cloud.stat", mean),
  deviance.funk = getOption("GeneFamilies.vector.cloud.deviance", sd))
}
\arguments{
\item{vecs.df}{A matrix of data.frame of row vectors of the same vector
space}

\item{stat}{A function used to infer the vector representing the argument
cloud. The function is applied on each column seperately in order to infer
the resulting vector's components. Default is
\code{getOption('GeneFamilies.vector.cloud.stat', mean)}.}

\item{deviance.funk}{Function applied on each column of the cloud
\code{vecs.df} in order to infer the cloud's spread around it's representing
vector. Default is \code{getOption('GeneFamilies.vector.cloud.deviance',
sd)}.}
}
\value{
A list with three named entries: \code{stat.vec} the vector
representing the cloud, \code{deviance.vec} the vector representing the
spread around \code{stat.vec}, and the vector orthogonal on the vector space
diagonal pointing to the \code{stat.vec}.
}
\description{
Computes the statistic \code{stat} (default is \code{base::mean}) vector of
the cloud of vectors. Also infers the deviance in terms of
\code{deviance.funk} (default is \code{base::sd}) of each axis of the vector
space.
}
