% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{subfunctionalizationAnnotationBased}
\alias{subfunctionalizationAnnotationBased}
\title{Identifies instances of subfunctionalization occurred in the evolution from
the base genes to the evolved genes. Subfunc. is identified based on the
gene annotations. All possible combinations of evolved genes are tested with
each base gene. Subfunc conditions are 1. The tested evolved genes
annotations (EA) are not a superset of the base gene annotations (BGA), 2.
the intersection of EA is smaller than the union of EA, and 3. The union of
EA is identical to the BGA, or, if not strict, the union of the EA is a
subset of the BGA.}
\usage{
subfunctionalizationAnnotationBased(base.genes, evolved.genes,
  annot.df = all.ipr, gene.col = 1, annot.col = 2, strict = TRUE)
}
\arguments{
\item{base.genes}{Character vector of gene IDs, identifying the base genes.}

\item{evolved.genes}{Character vector of gene IDs, identifying the evolved
genes.}

\item{annot.df}{A data.frame of gene annotations. Default is
\code{GeneFamilies::all.ipr}}

\item{gene.col}{The column name or index in \code{annot.df} in which to
lookup the gene IDs. Default is \code{1}}

\item{strict}{Boolean indicating whether to apply the strict criterium in
the second condition, or not (see above for details). Default is
\code{TRUE}.}

\item{anno.col}{The column name or index in \code{annot.df} in which to
lookup the gene annotations. Default is \code{2}.}
}
\value{
\code{NA} if there are no annotations for at least one base gene and
at least two evolved.genes. \code{NULL} is returned if the previous
conditions are met, but no subfunctionilizations could be identified, and
finally a data.frame is returned with the identified subfunctionilizations.
In this, columns are: base.gene, base.annotations, evolved.genes, and
evolved.annotations.
}
\description{
Identifies instances of subfunctionalization occurred in the evolution from
the base genes to the evolved genes. Subfunc. is identified based on the
gene annotations. All possible combinations of evolved genes are tested with
each base gene. Subfunc conditions are 1. The tested evolved genes
annotations (EA) are not a superset of the base gene annotations (BGA), 2.
the intersection of EA is smaller than the union of EA, and 3. The union of
EA is identical to the BGA, or, if not strict, the union of the EA is a
subset of the BGA.
}

