% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{subseqPos}
\alias{subseqPos}
\title{Infers the amino acid position(s) at which the argument 'aa.subseq' starts
in the full 'aa.full' amino acid sequence.}
\usage{
subseqPos(aa.subseq, aa.full)
}
\arguments{
\item{aa.subseq}{The amino acid subsequence to be matched}

\item{aa.full}{The full amino acid sequence}
}
\value{
An integer vector with matching start positions.
}
\description{
Infers the amino acid position(s) at which the argument 'aa.subseq' starts
in the full 'aa.full' amino acid sequence.
}
