% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{subTrees}
\alias{subTrees}
\title{Identifies a list of sub-trees in argument tree initiating from the the
argument nodes and maps those sub-trees on their contained tips.}
\usage{
subTrees(tree, nds = unique(tree$edge[, 1]))
}
\arguments{
\item{tree}{an instance of class ape::phylo}

\item{nds}{the inner nodes for which to identify the sub-trees}
}
\value{
A named list. Names are the argument tips and values the node label
of the root node of the sub-tree.
}
\description{
Identifies a list of sub-trees in argument tree initiating from the the
argument nodes and maps those sub-trees on their contained tips.
}
