% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{tandemsFromFamiliesAndNeighborhood}
\alias{tandemsFromFamiliesAndNeighborhood}
\title{Identifies tandem duplicated genes within a gene family. Conditions are that
the tandems belonging to the same species are within a gene neighborhod of
(default) nine genes on the same chromosome. See
\code{GeneFamilies::areNeighbors} for more details.
_Note_ that this approach does NOT exclude orthologous genes. If you want to
distinguish those do a subsequent \code{setdiff} with the result of this
function.}
\usage{
tandemsFromFamiliesAndNeighborhood(gene.family, spec.neighborhoods,
  spec.gene.ids.arg = spec.gene.ids)
}
\arguments{
\item{gene.family}{A character vector holding the gene identifiers of the
genes comprising a single gene family.}

\item{spec.neighborhoods}{A list with names representing the species and
values being the tables indicating gene neighborhood. See the Vignette
\code{GeneFamilies} on how to generate these neighborhood-tables. Also see
function \code{GeneFamilies::areNeighbors} for more details on these tables'
format.}

\item{spec.gene.ids.arg}{An instance of \code{base::list} with names being
the investigated species and values all protein accessions belonging to the
respective species. Default is \code{spec.gene.ids} (see file \code{zzz.R}
for more details).}
}
\value{
A character vector holding all Tandem Duplicated Genes found within
the argument gene family.
}
\description{
Identifies tandem duplicated genes within a gene family. Conditions are that
the tandems belonging to the same species are within a gene neighborhod of
(default) nine genes on the same chromosome. See
\code{GeneFamilies::areNeighbors} for more details.
_Note_ that this approach does NOT exclude orthologous genes. If you want to
distinguish those do a subsequent \code{setdiff} with the result of this
function.
}
