% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{testAlignedForUnalignedAAPos}
\alias{testAlignedForUnalignedAAPos}
\title{Test for function \code{alignedForUnalignedAAPos}.}
\usage{
testAlignedForUnalignedAAPos()
}
\value{
TRUE if and only if all tests pass.
}
\description{
Test for function \code{alignedForUnalignedAAPos}.
}
