% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{testExprVecSpaceEvolutionAfterDupl}
\alias{testExprVecSpaceEvolutionAfterDupl}
\title{Testing function \code{exprVecSpaceEvolutionAfterDupl}.}
\usage{
testExprVecSpaceEvolutionAfterDupl()
}
\value{
\code{TRUE} if and only if all tests pass
}
\description{
Testing function \code{exprVecSpaceEvolutionAfterDupl}.
}
