% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{testGetPairsFromDistForIndices}
\alias{testGetPairsFromDistForIndices}
\title{Tests the function \code{getPairsFromDistForIndices}.}
\usage{
testGetPairsFromDistForIndices()
}
\value{
\code{TRUE} if and only of all tests pass.
}
\description{
Tests the function \code{getPairsFromDistForIndices}.
}
