% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{testHomologs}
\alias{testHomologs}
\title{Test of function 'homologs(...)'}
\usage{
testHomologs()
}
\value{
TRUE if and only if the test succeeded.
}
\description{
Test of function 'homologs(...)'
}
