% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{testNonConservedAnnotations}
\alias{testNonConservedAnnotations}
\title{Test function \code{nonConservedAnnotations}}
\usage{
testNonConservedAnnotations()
}
\value{
TRUE if and only if all tests pass
}
\description{
Test function \code{nonConservedAnnotations}
}
