% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funks.R
\name{testNonConservedCompositeAnnotations}
\alias{testNonConservedCompositeAnnotations}
\title{Test for function nonConservedCompositeAnnotations.}
\usage{
testNonConservedCompositeAnnotations()
}
\value{
TRUE if and only if all tests pass
}
\description{
Test for function nonConservedCompositeAnnotations.
}
