% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationVectorSpace.R
\name{testSubfunctionalizationAnnotationBased}
\alias{testSubfunctionalizationAnnotationBased}
\title{Testing \code{subfunctionalizationAnnotationBased}}
\usage{
testSubfunctionalizationAnnotationBased()
}
\value{
\code{TRUE} if and only of all tests are passed successfully.
}
\description{
Testing \code{subfunctionalizationAnnotationBased}
}
