% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_funks.R
\name{testTandemsFromFamiliesAndNeighborhood}
\alias{testTandemsFromFamiliesAndNeighborhood}
\title{Function testing \code{GeneFamilies::tandemsFromFamiliesAndNeighborhood}.}
\usage{
testTandemsFromFamiliesAndNeighborhood()
}
\arguments{
\item{\code{NULL}}{}
}
\value{
TRUE if and only if all tests pass
}
\description{
Function testing \code{GeneFamilies::tandemsFromFamiliesAndNeighborhood}.
}
