% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toLowerCutTail}
\alias{toLowerCutTail}
\title{Function to sanitize maize genes. Cuts of the tail following \code{_} and
puts everything into lower case.}
\usage{
toLowerCutTail(x)
}
\arguments{
\item{x}{a character vector of maize gene accessions}
}
\value{
sanitized version of argument \code{x}
}
\description{
Function to sanitize maize genes. Cuts of the tail following \code{_} and
puts everything into lower case.
}

