% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{translate2AASeqs}
\alias{translate2AASeqs}
\title{Uses the program MACSE to translate the coding sequences to amino acid
sequences.}
\usage{
translate2AASeqs(path.2.cds.fasta, macse.call = getOption("macse.call",
  paste("java -Xmx600m -jar ", file.path(path.package("MaizeGeneFamilies"),
  "macse_v1.01b.jar"), " -prog translateNT2AA", sep = "")))
}
\arguments{
\item{path.2.cds.fasta}{gives the valid file path to the FASTA file holding
the coding sequences.}

\item{macse.call}{The call with which to invoke MACSE through the
system(...) function. Use option 'macse.call' to modify the default.}
}
\value{
TRUE if and only if no error has occurred.
}
\description{
Uses the program MACSE to translate the coding sequences to amino acid
sequences.
}

