% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{tree4Busted}
\alias{tree4Busted}
\title{Annotates a phylogenetic tree in such amanner that foreground branches are
recognized by HYPHY's BUSTED analysis.}
\usage{
tree4Busted(phylo.tree, foreground.nodes)
}
\arguments{
\item{phylo.tree}{an instance of class ape::phylo}

\item{foreground.nodes}{integer vector holding the nodes that are selected
as foreground}
}
\value{
A BUSTED annotated copy of 'phylo.tree'
Mark foreground tips / leaves:
Mark foreground inner nodes:
}
\description{
Annotates a phylogenetic tree in such amanner that foreground branches are
recognized by HYPHY's BUSTED analysis.
}
