% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{unalignedAAforAlignedAA}
\alias{unalignedAAforAlignedAA}
\title{Infers the corresponding un-aligned position of an selected amino acid
obtained from an multiple sequence alignment. This is done by iterativly
increasing a amino acid substring, without gap characters, and finding its
position on the original unaligned amino acid sequence.}
\usage{
unalignedAAforAlignedAA(gene, sel.aa, aa.fasta, msa.fasta, min.radius = 3)
}
\arguments{
\item{gene}{The gene accession / ID as used in both 'aa.fasta' and
'msa.fasta'}

\item{sel.aa}{The index of the homologous amino acid subject to selection
(integer coordinate)}

\item{aa.fasta}{The result of readAAStringSet(path_2_AAs.fasta) [package
Biostrings]}

\item{msa.fasta}{The result of
readAAMultipleAlignment(path_2_AAs_MSA.fasta) [package Biostrings]}

\item{min.radius}{The minimum radius with which to start selecting up- and
downstream subsequences}
}
\value{
An integer; either NA, if no position could be inferred, or the
corresponding un-aligned position.
}
\description{
Infers the corresponding un-aligned position of an selected amino acid
obtained from an multiple sequence alignment. This is done by iterativly
increasing a amino acid substring, without gap characters, and finding its
position on the original unaligned amino acid sequence.
}

