% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprSelectedPositions.R
\name{unalignedAAforAlignedAAPos}
\alias{unalignedAAforAlignedAAPos}
\title{Computes the position in the unaligned sequence of a character indicated by
\code{sel.aa}.}
\usage{
unalignedAAforAlignedAAPos(gene, sel.aa, msa.fasta,
  gap.char = getOption("GeneFamilies.gap.char", "-"),
  gap.char.matching.fixed = getOption("GeneFamilies.gap.char.matching.fixed",
  TRUE))
}
\arguments{
\item{gene}{The gene accession / ID as used in 'msa.fasta'}

\item{sel.aa}{The index of the homologous amino acid subject to selection
(integer coordinate)}

\item{msa.fasta}{The result of
\code{seqinr::read.fasta(path_2_AAs_MSA.fasta, seqtype='AA', as.string=TRUE,
strip.desc=TRUE)}.}

\item{gap.char}{The character or regular expression used to identify non
sequence characters in the aligned sequences. Default is
\code{getOption('GeneFamilies.gap.char', '-')}.}

\item{gap.char.matching.fixed}{boolean indicating the value to pass to
\code{grepl} and \code{gsub} argument \code{fixed=}. Set to \code{TRUE} if
\code{gap.char} is literal and atomic. Default is
getOption('GeneFamilies.gap.char.matching.fixed', TRUE).}
}
\value{
An integer; either NA if the position is a non sequence character,
or the corresponding un-aligned position.
}
\description{
Computes the position in the unaligned sequence of a character indicated by
\code{sel.aa}.
}

