% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{upperWhisker}
\alias{upperWhisker}
\title{Identifies the value that corresponds to the upper whisker end in an
boxplot. \code{NA} are ignored}
\usage{
upperWhisker(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
A single number
}
\description{
Identifies the value that corresponds to the upper whisker end in an
boxplot. \code{NA} are ignored
}
