% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{validate_data}
\alias{validate_data}
\title{Validate and Filter Loaded Data Objects}
\usage{
validate_data(loaded_objects, pattern)
}
\arguments{
\item{loaded_objects}{A character vector containing the names of objects loaded in the R environment.}

\item{pattern}{A character string representing the regex pattern to identify the names of the target data objects.}
}
\value{
A character vector of validated object names that match the specified pattern and contain valid data.
Invalid data types (e.g., non-lists and non-vectors), empty objects, or objects with only \code{NA} values are excluded.
}
\description{
This function filters a list of loaded data objects by a specified pattern and validates them to ensure they meet the required criteria.
It checks that each object is a non-empty list or vector and contains valid data (i.e., it is not entirely \code{NA} values).
}
\examples{
# Load data and then filter and validate the names of loaded gene expression profiles:
loaded_objects <- ls() # List all loaded objects
valid_data_names <- validate_data(loaded_objects, "(_dists$|\\\\.dists$)")
valid_data_names_tissue <- validate_data(loaded_objects, "(_dists_tissue$|\\\\.dists\\\\.tissue$)")

}
