% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies.R
\name{validateAASeqs}
\alias{validateAASeqs}
\title{Validates an AAString}
\usage{
validateAASeqs(aa.seq)
}
\arguments{
\item{'aa.seq'}{an instance of \code{seqinr::SeqFastaAA}}
}
\value{
TRUE if and only if the argument does not have a premature stop
codon.
}
\description{
Validates an AAString
}
