% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_funks.R
\name{vectorProjection}
\alias{vectorProjection}
\title{Computes the vector projection of vector \code{a} onto vector \code{b}, i.e.
the vector-segment of \code{b} 'overshadowed' by \code{a}. See
en.wikipedia.org/wiki/Vector_projection}
\usage{
vectorProjection(a, b)
}
\arguments{
\item{a}{numerical representing first vector}

\item{b}{numerical representing second vector}
}
\value{
The vector resulting from projecting vector \code{a} upon vector
\code{b}.
}
\description{
Computes the vector projection of vector \code{a} onto vector \code{b}, i.e.
the vector-segment of \code{b} 'overshadowed' by \code{a}. See
en.wikipedia.org/wiki/Vector_projection
}
