% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_funks.R
\name{writeAxt}
\alias{writeAxt}
\title{Writes pairwise aligned sequences to a file of AXT format for Ka/Ks
calculation.}
\usage{
writeAxt(axt, axt.path)
}
\arguments{
\item{axt}{a named list as returned from function 'axt(...)'}

\item{axt.path}{the valid file path indicating the output AXT file to
generate}
}
\value{
the result of calling readLines
}
\description{
Writes pairwise aligned sequences to a file of AXT format for Ka/Ks
calculation.
}
