% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funks.R
\name{calculate_exp.prof.dists.tissue.statistics}
\alias{calculate_exp.prof.dists.tissue.statistics}
\title{Calculate Statistics for Expression Profile Distances by Tissue}
\usage{
calculate_exp.prof.dists.tissue.statistics(data)
}
\arguments{
\item{data}{A nested list where each element represents a gene cluster. Each cluster contains a list of
tissues with associated expression profile distances. The structure should be such that each cluster (gene family)
contains named tissues as its sublist, and each tissue contains numeric distances.}
}
\value{
A tibble where each row corresponds to a gene cluster, and the columns include:
\itemize{
\item \code{Family}: The gene cluster name.
\item \code{Tissue}: The name of the tissue.
\item \code{Mean}: The mean of the expression profile distances for that tissue.
\item \code{Median}: The median of the expression profile distances for that tissue.
The result contains one row for each tissue in each gene cluster, with statistics for each tissue.
}
}
\description{
Computes the mean and median of the expression profile distances for each gene cluster,
per tissue within a given dataset. The output is structured such that tissues are represented in the columns,
with mean and median statistics for each tissue.
}
\examples{
# Calculate the mean and median of expression profile distances for each gene cluster per tissue
result <- calculate_exp.prof.dists.tissue.statistics(my_data)

}
